### 2025-02-24 #7840 @editorialbot
[PRE REVIEW]: Snowdrop:  Python Package for DSGE modeling
**Submitting author:** @agoumilevski  (<a href="http://orcid.org/0009-0004-5574-854X">Alexei Goumilevski </a>)
**Repository:** https://github.com/gumilevskij/Framework
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.0

Thanks for submitting your paper to JOSS @agoumilevski . **Currently, there isn't a JOSS editor assigned** to your paper.
