+++ pyproject.toml
@@ -1,9 +1,11 @@
 [tool.poetry]
-name = "snowdrop"
-version = "1.0.0"
+name = "pysnowdrop"
+version = "1.0.3"
 description = "Python Platform for DSGE Modeling"
-authors = ["Alexei Goumilevski <Agoumilevski@hotmail.com>",
+authors = ["Alexei Goumilevski <agoumilevski@hotmail.com>",
            "James Otterson <jjotterson@gmail.com>"]
+maintainers = ["Alexei Goumilevski <Agoumilevski@hotmail.com>",
+               "James Otterson <jjotterson@gmail.com>"]
 license = "LGPL-3.0-or-later"
 readme =  "snowdrop/docs/README.md"
 keywords = ["DSGE", "Macroeconomic Modeling"]
@@ -13,7 +15,6 @@ classifiers=[
     "Environment :: Console",
     "Intended Audience :: Science/Research",
     "Operating System :: OS Independent",
-    "Topic :: Macroeconomic Modeling :: DSGE",
     "Topic :: Software Development"
 ]
 
@@ -29,6 +30,8 @@ include = [
     { path = "snowdrop/api_docs", format = ["sdist", "wheel"] }
 ]
 
+exclude = ["snowdrop/src/troll", "snowdrop/data/IMFE"]
+
 [tool.poetry.scripts]
 test = "snowdrop.src.tests.test_toy_models:test"
 
@@ -39,9 +42,9 @@ Repository = "https://github.com/gumilevskij/Framework.git"
 [tool.poetry.dependencies]
 python =  "^3.10"
 pandas =  "^2.1.4"
-scipy =  "^1.11.4"
+scipy =  "1.12.0"
 numpy =  "^1.26.4"
-matplotlib =  "^3.8.0"
+matplotlib =  "3.10.0"
 sympy =  "^1.12"
 symengine =  "^0.11.0"
 emcee =  "^3.1.6"
@@ -61,6 +64,8 @@ pytest =  "^7.4.0"
 pykalman  =  "^0.9.7"
 pylatex =  "^1.4.2"
 pdfkit = "^1.0.0"
+xlsxwriter = "^3.2.0"
+"ruamel.yaml" = "0.16.3"
 
 # List of optional dependencies
 compecon = {version="^2024.5.19", optional=true}
@@ -68,7 +73,6 @@ quantecon = {version="^0.7.2", optional=true}
 reportlab = {version="^4.2.2", optional=true}
 textwrap3 = {version="^0.9.2", optional=true}
 PyLaTeX = {version="^1.4.2", optional=true}
-"ruamel.yaml" = {version="^0.18.6", optional=true}
 glob2 = {version="^0.7", optional=true}
 autograd = {version="^1.7.0", optional=true}
 pdflatex =  {version="^0.1.3", optional=true}
@@ -79,7 +83,6 @@ scikit-learn =  ["scikit-learn"]
 pylatex =  ["PyLaTeX"]
 textwrap3 =  ["textwrap3"]
 glob2 =  ["glob2"]
-"ruamel.yaml" =  ["ruamel.yaml"]
 pykalman  = ["pykalman"]
 compecon = ["compecon"]
 quantecon = ["quantecon"]
