### 2025-02-25 c3dca5e
Update linter.py
