+++ joss_submission/paper.bib
@@ -1,5 +1,5 @@
 @article{Eichenbaum, 
-  author = {Eichenbaum Martin, Rebelo Sergio, Trabandt Mathias},
+  author = {Eichenbaum Martin and Rebelo Sergio and Trabandt Mathias},
   journal = {NBER Working Paper, No. 26882},
   title = {The Macroeconomics of Epidemics},
   year = 2020
@@ -49,21 +49,21 @@
 }
 
 @article{Juillard, 
-  author = {Juillard Michel, Laxton Douglas, McAdam Peter, Pioro Hope},
+  author = {Juillard Michel and Laxton Douglas and McAdam Peter and Pioro Hope},
   journal = {Journal of Economic Dynamics and Control, No. 22, pp.1291—1318},
   title = {An algorithm competition: First-order iterations versus Newton-based technique},
   year = 1998
 }
 
 @article{Goumilevski, 
-  author = {Goumilevski Alexei, Farias Aquiles, Tanyeri Kadir},
+  author = {Goumilevski Alexei and Farias Aquiles and Tanyeri Kadir},
   journal = {Proceedings of the 27th International Conference on Computing in Economics and Finance},
   title = {A New Macroeconomic Modeling Framework Applied to Assess Effects of COVID-19},
   year = 2021
 }
 
 @article{Ireland, 
-  author = {Ireland Peter},
+  author = {Peter Ireland},
   journal = {NBER Working Paper, No. 10309},
   title = {Technology Shocks in the New Keynesian Model},
   year = 2004
