+++ joss_submission/paper.bib
@@ -1,5 +1,5 @@
 @article{Eichenbaum, 
-  author = {Eichenbaum Martin and Rebelo Sergio and Trabandt Mathias},
+  author = {Martin Eichenbaum and Sergio Rebelo and Mathias Trabandt},
   journal = {NBER Working Paper, No. 26882},
   title = {The Macroeconomics of Epidemics},
   year = 2020
@@ -49,14 +49,14 @@
 }
 
 @article{Juillard, 
-  author = {Juillard Michel and Laxton Douglas and McAdam Peter and Pioro Hope},
+  author = {Michel Juillard and Douglas Laxton and Peter McAdam and Hope Pioro},
   journal = {Journal of Economic Dynamics and Control, No. 22, pp.1291—1318},
   title = {An algorithm competition: First-order iterations versus Newton-based technique},
   year = 1998
 }
 
 @article{Goumilevski, 
-  author = {Goumilevski Alexei and Farias Aquiles and Tanyeri Kadir},
+  author = {Alexei Goumilevski and Aquiles Farias and Kadir Tanyeri},
   journal = {Proceedings of the 27th International Conference on Computing in Economics and Finance},
   title = {A New Macroeconomic Modeling Framework Applied to Assess Effects of COVID-19},
   year = 2021
