+++ snowdrop/__init__.py
@@ -1,6 +1,6 @@
 import os,sys
 path = os.path.dirname(os.path.abspath(__file__))
-working_dir = os.path.abspath(path + "/../../..")
+working_dir = os.path.abspath(path + "/..")
 sys.path.append(working_dir)
 os.chdir(working_dir)
 
