### 2025-02-27 915ef54
Update linter.py
