### 2025-02-27 79a01f3
Delete snowdrop directory
