+++ snowdrop/src/numeric/optimization/__init__.py
@@ -1,4 +1,0 @@
-__all__ = ["optimize","util"]
-
-
-
