+++ snowdrop/src/numeric/solver/tunes.py
@@ -1,1155 +1,0 @@
-"""
-Implements judgemental adjustments to endogenous variables paths.
-
-Created on Fri Oct 23, 2020
-Part of the code translated from Matlab Iris to Python.
-
-@author: A.Goumlevski
-"""
-import numpy as np
-import scipy.linalg as la
-itr = 0
-
-def forecast_with_tunes(model,Nperiods=0,y=None,T=None,Re=None,C=None,shocks=None,has_imaginary_shocks=False,has_imaginary_values=False,scale=complex(1,0)):
-    """
-    Forecast with judgmental adjustments (conditional forecasts).
-    This is a solution for a 'fixed path' of endogenous variables.
-    
-    For references please see Michal Andrle's algorithm described in an unpubished paper:     
-    https://michalandrle.weebly.com/uploads/1/3/9/2/13921270/iris_simulate.pdf
-    
-    
-    Parameters:
-        model : Model.
-            Model object.
-        Nperiods : numpy.ndarray.
-            Number of time periods.
-        y : numpy.ndarray.
-            The solution with no tunes.
-        T : numpy.ndarray.
-            Transition matrix.
-        Re : numpy.ndarray.
-            Matrix of coefficiets of forward shocks.
-        C : numpy.ndarray.
-            Matrix of constants.
-        shocks : list.
-            Shocks at different time periods.
-        has_imaginary_shocks : bool.
-            Flag set to True if ay of the shocks is a complex number, and to False other otherwise.
-        has_imaginary_values : bool.
-            Flag set to True if any of the endogenous variables path is a complex number, and to False other otherwise.
-        scale : numpy.complex.
-            Scaling factor for standard deviations of expected and un-expected shocks.
-            If model.anticipate flag is set to True , then real part of scale is a multiplier
-            for standard devitions of expected shocks and the imaginary part is a multiplier
-            for standard devitions of un-expected shocks.
-            If model.anticipate flag is set to False, then it is wise versa.  
-           
-    Returns:
-        model : Model.
-            Model object.
-        y : numpy array.
-            Numerical solution with tunes.
-            
-    """
-    if model.isLinear:
-       y,adjusted_shocks = mixed(model=model,x=y,T=T,Re=Re,C=C,shocks=shocks,Nperiods=Nperiods,scale=scale)
-       if "periods" in model.options:
-           del model.options["periods"]
-       return y,adjusted_shocks[1:],model
-    else:
-       y,adjusted_shocks,err,elapsed = non_linear_mixed(model=model,shocks=shocks,Nperiods=Nperiods)
-       return y,adjusted_shocks[1:],model,itr,err,elapsed 
-      
-
-def mixed(model,x,T,Re,C,shocks,Nperiods,scale=complex(1,0)):
-    """
-    Forecast with judgmental adjustments (conditional forecasts).
-    
-    This is a generic case with a mixture of anticipated and un-anticipated shocks.  
-    This soltion also can handle conditional shocks.
-    
-    If model anticipate flag is set to True, then real numbers describe anticipated shocks
-    and imaginary numbers describe un-anticipated shocks.
-    If model anticipate flag is set to False, then real numbers describe un-anticipated shocks
-    and imaginary numbers describe anticipated shocks.
-    
-    Parameters:
-        model : Model.
-            `Model' object.
-        x : numpy.ndarray.
[…]
