+++ snowdrop/src/numeric/dp/test.py
@@ -1,600 +1,0 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-"""
-Created on Mon Feb  8 17:50:30 2021
-
-Originally developed by Fabrice Collard.
-
-Please see Dynamic Programming Notes #7 at:
-http://fabcol.free.fr/pdf/lectnotes7.pdf
-
-Translated from Matlab code to Python by A.Gouilevski
-"""
-
-import os, math
-from time import time
-import numpy as np
-from numpy.matlib import repmat
-from scipy.optimize import fmin
-from scipy.interpolate import CubicSpline
-import scipy.sparse as sparse
-from scipy.linalg import pinv
-from scipy.sparse.linalg import spsolve
-from snowdrop.src.numeric.dp.util import Chebychev_Polinomial, tv, scalar_tv, markov
-from snowdrop.src.graphs.util import plot
-
-sigma   = 1.50                     # utility parameter
-delta   = 0.10                     # depreciation rate
-beta    = 0.95                     # discount factor
-alpha   = 0.30                     # capital elasticity of output
-
-nbk     = 100                      # number of data points in the grid
-nbc     = 10
-nba     = 2
-crit    = 1                        # convergence criterion
-itr     = 0                        # iteration
-epsi    = 1e-6                     # convergence tolerance
-
-ks      = ((1-beta*(1-delta))/(alpha*beta))**(1/(alpha-1))
-dev     = 0.9                      # maximal deviation from steady state
-kmin    = (1-dev)*ks               # lower bound on the grid
-kmax    = (1+dev)*ks               # upper bound on the grid
-devk    = (kmax-kmin)/(nbk-1)      # implied increment
-kgrid   = np.linspace(kmin,kmax,nbk)  # build the grid
-v       = np.zeros(nbk)            # value function
-kp0     = 0                        # initial guess on k(t+1)
-dr      = np.zeros(nbk,dtype=int)  # decision rule (will contain indices)
-
-cmin    = 0.01                     # lower bound on the grid
-cmax    = kmax**alpha              # upper bound on the grid
-c       = np.linspace(cmin,cmax,nbc)  # build the grid
-Tv      = np.zeros(nbk)            # value function
-
-u       = (c**(1-sigma)-1)/(1-sigma)
-k		= np.linspace(kmin,kmax,nbk)
-
-
-def value_iteration(interpolate=False):
-    """
-    Solve the OGM by value function iteration.
-
-    Parameters:
-        interpolate : bool, optional
-            If True performs interpolation of value function to grid nodes. 
-            The default is False.
-
-    """
-    global itr, crit, nbk, kp0, kgrid, v, dr
-    t0 = time()
-    
-    if interpolate:
-        
-        nbc = 1000; nbk = 100
-        c     = np.linspace(cmin,cmax,nbc)  # builds the grid
-        kgrid = np.linspace(kmin,kmax,nbk)  # builds the grid
-        Tv    = np.zeros(nbk)               # value function
-        csy	  = 1-alpha*beta*delta/(1-beta*(1-delta))
-        v     = ((csy*kgrid**alpha)**(1-sigma)-1)/(1-sigma) # value function
-        u     = (c**(1-sigma)-1)/(1-sigma)
-        dr    = np.zeros(nbk,dtype=int) 
-        
-        ### Main loop
-        while crit>epsi:
-            for i in range(nbk):
-                 kp  = kgrid[i]**alpha+(1-delta)*kgrid[i]-c
[…]
