+++ snowdrop/src/numeric/grids/__init__.py
@@ -1,4 +1,0 @@
-__all__ = ["grid"]
-
-
-
