+++ snowdrop/models/MPAF/model.tex
@@ -1,0 +1,129 @@
+\documentclass{article}%
+\usepackage[T1]{fontenc}%
+\usepackage[utf8]{inputenc}%
+\usepackage{lmodern}%
+\usepackage{textcomp}%
+\usepackage{lastpage}%
+%
+\title{\textbf{Model File}}%
+\author{\textit{Generated by Python Framework}}%
+\date{\today}%
+%
+\begin{document}%
+\normalsize%
+\maketitle%
+\section{Model Information}%
+\label{sec:ModelInformation}%
+name: %
+\textit{Basic Quarterly Projection Model (QPM)}%
+\newline%
+file: %
+\textit{/home/alexei/work/Framework/examples/models/ICD/MPAF/model.yaml}%
+\subsection{Endogenous Variables Initial Values}%
+\label{subsec:EndogenousVariablesInitialValues}%
+D4L\_CPI = nan, D4L\_CPI\_TAR = nan, D4L\_GDP = nan, D4L\_S = nan, DLA\_CPI = nan, DLA\_CPI\_RW = nan, DLA\_GDP = nan, DLA\_GDP\_BAR = nan, DLA\_S = nan, DLA\_Z = nan, DLA\_Z\_BAR = nan, E\_DLA\_CPI = nan, L\_CPI = nan, L\_CPI\_RW = nan, L\_GDP = nan, L\_GDP\_BAR = nan, L\_GDP\_GAP = nan, L\_GDP\_RW\_GAP = nan, L\_S = nan, L\_Z = nan, L\_Z\_BAR = nan, L\_Z\_GAP = nan, MCI = nan, PREM = nan, RMC = nan, RR = nan, RR\_BAR = nan, RR\_GAP = nan, RR\_RW = nan, RR\_RW\_BAR = nan, RR\_RW\_GAP = nan, RS = nan, RSNEUTRAL = nan, RS\_RW = nan
+
+%
+\subsection{Measurement Variables}%
+\label{subsec:MeasurementVariables}%
+OBS\_D4L\_CPI\_TAR, OBS\_DLA\_CPI\_RW, OBS\_L\_CPI, OBS\_L\_GDP, OBS\_L\_GDP\_RW\_GAP, OBS\_L\_S, OBS\_RS, OBS\_RS\_RW
+
+%
+\subsection{Parameters}%
+\label{subsec:Parameters}%
[…]
