+++ snowdrop/__init__.py
@@ -1,0 +1,7 @@
+import os,sys
+path = os.path.dirname(os.path.abspath(__file__))
+working_dir = os.path.abspath(path + "/..")
+sys.path.append(working_dir)
+os.chdir(working_dir)
+
+
