+++ snowdrop/src/samples/makedata.py
@@ -1,0 +1,235 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Tue Apr 23 14:12:20 2019
+
+@author: Alexei Goumilevski
+
+ Reproduces results of  https://www.edx.org Monetary Policy Analysis and Forecasting course.
+"""
+import os
+import pandas as pd
+import numpy as np
+from math import ceil
+import matplotlib.pyplot as plt
+from datetime import datetime as dt
+import ruamel.yaml as yaml
+
+working_dir = os.path.abspath(os.path.join(os.path.dirname(__file__),"../../.."))
+os.chdir(working_dir)
+
+from snowdrop.src.numeric.filters.filters import HPF as hpf
+from snowdrop.src.numeric.filters.filters import BPASS as bpass
+from snowdrop.src.numeric.filters.filters import LRXfilter as lrx
+from snowdrop.src.numeric.sa.x13 import x13
+from snowdrop.src.graphs.util import plotTimeSeries
+from snowdrop.src.utils.load import read_file_or_url
+ 
+
+bin_dir = os.path.abspath(os.path.join(working_dir,"bin"))
+if os.path.exists(bin_dir):
+    sys_path = os.environ['PATH'] 
+    if not bin_dir in sys_path:
+        os.environ['PATH'] += ":" + bin_dir
+    os.environ['X13PATH'] = bin_dir
+
+    
+def makedata(Plot=False,save=False):
+    """
+    Prepares data.
+    """
+    plt.close('all')
+    
+    path_to_dir = os.path.abspath(os.path.join(working_dir,"graphs"))
+    
+    ## Load quarterly data
+    file_path = os.path.abspath(os.path.join(working_dir, "snowdrop/data/MPAF/data.csv"))
+    df = pd.read_csv(filepath_or_buffer=file_path,sep=',',header=0,index_col=0,parse_dates=True,infer_datetime_format=True)
+    df.index = pd.to_datetime(df.index,format="#Q")  # Quartely data frequency
+    
+    # Load quarterly data
+    d = {} 
+    for c in df.columns:
+        d[c] = df[c]
+          
+    ## Seasonal adjustment
+    count = 0
+    lst = []
+    for name in df.columns:
+        if len(name)>1:
+            if '_U' == name[-2:]:
+                count += 1
+                series = df[name].dropna()
+                df_adj = x13(series=series,freq='Q')[0]
+                d[name[:-2]] = df_adj["Adjusted"]
+                lst.append((series,df_adj))
+                
+    ## Make log of variables
+    exceptions = ['RS','RS_RW','D4L_CPI_TAR']
+    m = {}
+    for n in d:
+        if not n in exceptions:
+            m['L_'+n] = 100*np.log(d[n])
+    d = {**d, **m}
+
+    ##  Define the real exchange rate
+    d["L_Z"] = d["L_S"] + d["L_CPI_RW"] - d["L_CPI"]
+
+    ## Growth rate qoq, yoy
+    m = {}
+    for n in d:
+        if n.startswith('L_'):
+            m['DLA_'+n[2:]] = 4*d[n].diff(periods=1)
+            m['D4L_'+n[2:]] = d[n].diff(periods=4)
+    d = {**d, **m}
+
+    ##  Real variables
+    ##  Domestic real interest rate
+    d["RR"] = d["RS"] - d["D4L_CPI"]
+
+    ##Foreign real interest rate
+    d["RR_RW"] = d["RS_RW"] - d["D4L_CPI_RW"]
+
+    ## Trends and Gaps - Hodrick-Prescott filter
+    ## Even seasonally adjusted HDP is very volatile
+    var = ['RR','L_Z','RR_RW'];
+    m = {}
[…]
