+++ joss_submission/paper.bib
@@ -2,7 +2,9 @@
   author = {Martin Eichenbaum and Sergio Rebelo and Mathias Trabandt},
   journal = {NBER Working Paper, No. 26882},
   title = {The Macroeconomics of Epidemics},
-  year = 2020
+  year = 2020,
+  doi = {10.3386/w26882},
+  url = {https://www.nber.org/papers/w26882}
 }
 
 @article{fernandez,
@@ -13,14 +15,17 @@
   number={1},
   pages={229--252},
   year={2021},
-  publisher={Annual Reviews}
+  publisher={Annual Reviews},
+  doi = {10.3386/w27715},
+  url = {https://www.nber.org/papers/w27715}
 }
 
 @article{botman,
   title={DSGE modeling at the fund: Applications and further developments},
   author={Botman, Dennis PJ and Karam, Philippe D and Laxton, Douglas and Rose, David},
   year={2007},
-  publisher={IMF Working Paper}
+  publisher={IMF Working Paper},
+  url = {https://www.imf.org/en/Publications/WP/Issues/2016/12/31/DSGE-Modeling-at-the-Fund-Applications-and-Further-Developments-21265}
 }
 
 @article{smets,
@@ -30,7 +35,9 @@
   volume={1},
   pages={51--65},
   year={2010},
-  publisher={Springer}
+  publisher={Springer},
+  doi = {10.1007/s13209-010-0020-9},
+  url = {https://d-nb.info/1100877037/34}
 }
 
 @article{del,
@@ -38,33 +45,39 @@
   author={Del Negro, Marco and Eusepi, Stefano and Giannoni, Marc P and Sbordone, Argia M and Tambalotti, Andrea and Cocci, Matthew and Hasegawa, Raiden and Linder, M},
   journal={FRB of New York Staff Report},
   number={647},
-  year={2013}
+  year={2013},
+  url = {https://www.newyorkfed.org/medialibrary/media/research/staff_reports/sr647.pdf}
 }
 
 @article{yagihashi,
   title={DSGE models used by policymakers: A survey},
   author={Yagihashi, Takeshi},
   journal={Policy Research Institute, Ministry of Finance, Japan, PRI Discussion Paper Series},
-  year={2020}
+  year={2020},
+  url = {https://www.mof.go.jp/pri/research/discussion_paper/ron333.pdf}
 }
 
 @article{Juillard, 
   author = {Michel Juillard and Douglas Laxton and Peter McAdam and Hope Pioro},
   journal = {Journal of Economic Dynamics and Control, No. 22, pp.1291—1318},
   title = {An algorithm competition: First-order iterations versus Newton-based technique},
-  year = 1998
+  year = 1998,
+  doi = {10.1016/S0165-1889(98)00013-X},
+  url = {https://www.sciencedirect.com/science/article/abs/pii/S016518899800013X}
 }
 
 @article{Goumilevski, 
   author = {Alexei Goumilevski and Aquiles Farias and Kadir Tanyeri},
   journal = {Proceedings of the 27th International Conference on Computing in Economics and Finance},
   title = {A New Macroeconomic Modeling Framework Applied to Assess Effects of COVID-19},
-  year = 2021
+  year = 2021,
+  url = {https://www.academia.edu/117687479/A_New_Macroeconomic_Modeling_Platform_Applied_to_Assess_Effects_of_COVID_19}
 }
 
 @article{Ireland, 
   author = {Peter Ireland},
   journal = {NBER Working Paper, No. 10309},
   title = {Technology Shocks in the New Keynesian Model},
-  year = 2004
+  year = 2004,
+  url = {https://www.nber.org/system/files/working_papers/w10309/w10309.pdf}
 }
