+++ joss_submission/paper.bib
@@ -8,51 +8,51 @@
 }
 
 @article{fernandez,
-  title={Estimating DSGE models: Recent advances and future challenges},
-  author={Fern{\'a}ndez-Villaverde, Jes{\'u}s and Guerr{\'o}n-Quintana, Pablo A},
-  journal={Annual Review of Economics},
-  volume={13},
-  number={1},
-  pages={229--252},
-  year={2021},
-  publisher={Annual Reviews},
+  title = {Estimating DSGE models: Recent advances and future challenges},
+  author = {Fern{\'a}ndez-Villaverde, Jes{\'u}s and Guerr{\'o}n-Quintana, Pablo A},
+  journal = {Annual Review of Economics},
+  volume = {13},
+  number = {1},
+  pages = {229--252},
+  year = {2021},
+  publisher = {Annual Reviews},
   doi = {10.3386/w27715},
   url = {https://www.nber.org/papers/w27715}
 }
 
 @article{botman,
-  title={DSGE modeling at the fund: Applications and further developments},
-  author={Botman, Dennis PJ and Karam, Philippe D and Laxton, Douglas and Rose, David},
-  year={2007},
-  publisher={IMF Working Paper},
+  title = {DSGE modeling at the fund: Applications and further developments},
+  author = {Botman, Dennis PJ and Karam, Philippe D and Laxton, Douglas and Rose, David},
+  year = {2007},
+  publisher = {IMF Working Paper},
   url = {https://www.imf.org/en/Publications/WP/Issues/2016/12/31/DSGE-Modeling-at-the-Fund-Applications-and-Further-Developments-21265}
 }
 
 @article{smets,
-  title={DSGE models and their use at the ECB},
-  author={Smets, Frank and Christoffel, Kai and Coenen, G{\"u}nter and Motto, Roberto and Rostagno, Massimo},
-  journal={SERIEs},
-  volume={1},
-  pages={51--65},
-  year={2010},
-  publisher={Springer},
+  title = {DSGE models and their use at the ECB},
+  author = {Smets, Frank and Christoffel, Kai and Coenen, G{\"u}nter and Motto, Roberto and Rostagno, Massimo},
+  journal = {SERIEs},
+  volume = {1},
+  pages = {51--65},
+  year = {2010},
+  publisher = {Springer},
   doi = {10.1007/s13209-010-0020-9},
   url = {https://d-nb.info/1100877037/34}
 }
 
 @article{del,
-  title={The FRBNY DSGE model},
-  author={Del Negro, Marco and Eusepi, Stefano and Giannoni, Marc P and Sbordone, Argia M and Tambalotti, Andrea and Cocci, Matthew and Hasegawa, Raiden and Linder, M},
-  journal={FRB of New York Staff Report},
-  number={647},
-  year={2013},
+  title = {The FRBNY DSGE model},
+  author = {Del Negro, Marco and Eusepi, Stefano and Giannoni, Marc P and Sbordone, Argia M and Tambalotti, Andrea and Cocci, Matthew and Hasegawa, Raiden and Linder, M},
+  journal = {FRB of New York Staff Report},
+  number = {647},
+  year = {2013},
   url = {https://www.newyorkfed.org/medialibrary/media/research/staff_reports/sr647.pdf}
 }
 
 @article{yagihashi,
-  title={DSGE models used by policymakers: A survey},
-  author={Yagihashi, Takeshi},
-  journal={Policy Research Institute, Ministry of Finance, Japan, PRI Discussion Paper Series},
+  title = {DSGE models used by policymakers: A survey},
+  author = {Yagihashi, Takeshi},
+  journal = {Policy Research Institute, Ministry of Finance, Japan, PRI Discussion Paper Series},
[…]
