+++ snowdrop/src/numeric/filters/filters.py
@@ -140,7 +140,7 @@ def HPfilter(data, lmbd=1600):
         
     .. math::
         
-        \min_{T_{t}}[ \sum_{t=0}((X_{t} - T_{t})^2 + \lambda*((T_{t+1} - T_{t}) - (T_{t} - T_{t-1}))^2) ]
+        min_{T_{t}}[ \sum_{t=0}((X_{t} - T_{t})^2 + \lambda*((T_{t+1} - T_{t}) - (T_{t} - T_{t-1}))^2) ]
 
 
     Args:
