### 2025-04-18 fa264e9
Update test_toy_models.py
