+++ snowdrop/src/tests/test_toy_models.py
@@ -43,53 +43,57 @@ def test_2(fname='models/TOY/JLMP98.yaml',decompose=True,title=None):
     # Create model object
     model = driver.importModel(fname=file_path,use_cache=False)
     
-    # # Solving non-linear model by linear solver.  Just kidding...
-    # from snowdrop.src.utils.equations import topology
-    # model.isLinear = True
-    # topology(model) 
+    decomp = None;  output_variables = None
+    
+    if 'JLMP98' in fname:
+        
+        # # Solving non-linear model by linear solver.  Just kidding...
+        # from snowdrop.src.utils.equations import topology
+        # model.isLinear = True
+        # topology(model) 
 
-    variables = model.symbols['variables']
-    # Change model parameters
-    cal = {'g':0.049,'p_pdot1':0.414,'p_pdot2': 0.196,'p_pdot3': 0.276,
-            'p_rs1':3,'p_y1':0.304,'p_y2':0.098,'p_y3':0.315}
-    model.calibration['parameters'] = list(cal.values())
+        variables = model.symbols['variables']
+        # Change model parameters
+        cal = {'g':0.049,'p_pdot1':0.414,'p_pdot2': 0.196,'p_pdot3': 0.276,
+                'p_rs1':3,'p_y1':0.304,'p_y2':0.098,'p_y3':0.315}
+        model.calibration['parameters'] = list(cal.values())
+        
+        # Shocks
+        model.options["periods"] = [1]
+        model.options["shock_values"] = [0.02]
     
-    # Shocks
-    model.options["periods"] = [1]
-    model.options["shock_values"] = [0.02]
-    
-    # # User tune
-    # m = {'PDOT': pd.Series([0.01,0.01],[2,3])}
-    # shock_names = ['epdot']
-    # model.swap(var1=m,var2=shock_names,reset=False)
-    
-    # Exogenous variables. Revision of Monetary Policy Rate.
-    # The last exogenous process value is set for the rest of time periods
-    exo = pd.Series([0.03,0.0],[4,5])
-    exog_data = {'exo': exo}
-    model.symbolic.exog_data = exog_data     
-    model.calibration["exogenous"] = getExogenousSeries(model)
-    
-    # List of variables for which decomposition plots are produced
-    decomp = ['PDOT','RR','RS','Y']
-    output_variables = None #['PDOT','RR','RS','Y','PIE','LGDP','G','L_GDP','L_GDP_GAP']
-    if not decompose:
-        output_variables,decomp = decomp,output_variables
+        # # User tune
+        # m = {'PDOT': pd.Series([0.01,0.01],[2,3])}
+        # shock_names = ['epdot']
+        # model.swap(var1=m,var2=shock_names,reset=False)
+        
+        # Exogenous variables. Revision of Monetary Policy Rate.
+        # The last exogenous process value is set for the rest of time periods
+        # exo = pd.Series([0.03,0.0],[4,5])
+        # exog_data = {'exo': exo}
+        # model.symbolic.exog_data = exog_data     
+        # model.calibration["exogenous"] = getExogenousSeries(model)
+        
+        # List of variables for which decomposition plots are produced
+        decomp = ['PDOT','RR','RS','Y']
+        output_variables = None #['PDOT','RR','RS','Y','PIE','LGDP','G','L_GDP','L_GDP_GAP']
+        if not decompose:
+            output_variables,decomp = decomp,output_variables
     
     # Function that runs simulations, model parameters estimation, MCMC sampling, etc...
     y,dates = driver.run(model=model,decomp_variables=decomp,
                          output_variables=output_variables, #Solver="ABLR",
                          header=title,Output=True,Plot=True)
+    if 'JLMP98' in fname:
[…]
