### 2025-04-24 ac10be6
Delete snowdrop/models/MPAF/model.tex
