+++ snowdrop/models/MPAF/model.yaml.txt
@@ -1,339 +1,0 @@
-name:   Basic Quarterly Projection Model (QPM)   
-
-#Definitions:
-#   Transition_variables
-#   Real GDP (100*log)                                               L_GDP
-#   Trend in Real GDP (100*log)                                      L_GDP_BAR
-#   Output Gap (in      #)                                           L_GDP_GAP
-#   Quarterly Growth in Real GDP(in      # pa)                       DLA_GDP
-#   Real GDP Growth YoY (in      # pa)                               D4L_GDP
-#   Real GDP Trend Growth QoQ annualized (in      # pa)              DLA_GDP_BAR
-
-#   Real Monetary Condition Index (in      # pa)                     MCI
-
-#   CPI (level, 100*log)                                             L_CPI
-#   CPI Inflation QoQ annualized (in      # pa)                      DLA_CPI
-#   Expected CPI Inflation QoQ annualized (in      # pa)             E_DLA_CPI
-#   CPI Inflation YoY (in      # pa)                                 D4L_CPI
-#   Inflation Target (in      # pa)                                  D4L_CPI_TAR
-
-#   Real Marginal Cost  (in      #)                                  RMC
-
-#   Nominal Exchange Rate (LCY/FCY, 100*log)                         L_S
-#   Nominal Exch. Rate Depreciation QoQ annualized (in      # pa)    DLA_S
-#   Nominal Exch. Rate Depreciation YoY (in      # pa)               D4L_S
-#   Country Risk Premium (in      # pa)                              PREM
-
-#   Nominal Policy Interest Rate (in      # pa)                      RS
-#   Real Interest Rate (in      # pa)                                RR
-#   Trend Real Interest Rate (in      # pa)                          RR_BAR
-#   Real Interest Rate Gap (in      #)                               RR_GAP
-#   Nominal Policy Neutral Interest Rate (in      # pa)              RSNEUTRAL
-
-#   Real Exchange Rate (level, 100*log)                              L_Z
-#   Trend Real Exchange Rate (level, 100*log)                        L_Z_BAR
-#   Real Exchange Rate Gap (in      #)                               L_Z_GAP
-#   Real Exchange Rate Depreciation QoQ annualized (in      # pa)    DLA_Z
-#   Trend Real Exchange Rate Depreciation QoQ annualized(in # pa)    DLA_Z_BAR
-
-#   Foreign Output Gap (in      #)                                   L_GDP_RW_GAP
-#   Foreign Nominal Interest Rate (in      # pa)                     RS_RW
-#   Foreign Real Interest Rate (in      # pa)                        RR_RW 
-#   Foreign Real Interest Rate Trend (in      # pa)                  RR_RW_BAR
[…]
