### 2025-04-24 0c7a5b1
Delete snowdrop/models/COVID19/ert1_model.yaml.txt
