+++ snowdrop/models/COVID19/ert1_model.yaml.txt
@@ -1,517 +1,0 @@
-name: Eichenbaum, Rebelo and Trabandt Model with Resistant Virus Strain.  
-
-# This file solves and simulates the model developed by 
-# Eichenbaum M.S., Rebelo S., Trabandt M. (2020) “Epidemics in the Neoclassical and New Keynesian Models”.
-# NBER Working Paper 27430, http://www.nber.org/papers/w27430
-
-# Multi-strain coinfection model is borrowed from paper of 
-# S.Y. Tchoumi, H. Rwezaura, J.M. Tchuenche
-# Dynamic of a two-strain COVID-19 model with vaccination
-# https://assets.researchsquare.com/files/rs-553546/v1_covered.pdf?c=1631869530
-
-# Based on parameters, the code below nests the models with 
-# perfect competition (gam -> 1, xi = 0), imperfect competition with 
-# flexible prices (gam = 1.35, xi = 0) and imperfect competition with 
-# sticky prices (gam = 1.35, xi = 0.98 (weekly calibration)).
-# gam and xi are set in lines 56 and 60.
-
-# We solve the fully nonlinear model. Given the strong nonlinearities of the 
-# model due to the epidemic dynamics we use homotopy for a number of 
-# parameters, i.e. we increase parameter values stepwise until we impose
-# their final desired value. Homotopy calculations can take a while. 
-
-
-symbols:
-
-  variables: [
-    # Endogenous variables of the actual (sticky price) economy
-    y,k,n,w,rk,x,c,s,s1,s2,i,i1,i2,r,r1,r2,v,ns,ni,nr,cs,ci,cr,tau,tau1,tau2,
-    lambtilde,lamtau,lami,lams,lamr,dd,pop,Rb,pie,mc,F,Kf,rr,pbreve,
-    # Endogenous variables of the flexible price economy
-    yF,kF,nF,wF,rkF,xF,cF,sF,sF1,sF2,iF,iF1,iF2,rF,rF1,rF2,vF,nsF,niF,nrF,csF,ciF,crF,
-    tauF,tauF1,tauF2,lambtildeF,lamtauF,lamiF,lamsF,lamrF,ddF,popF,RbF,pieF,
-    mcF,FF,KfF,rrF,pbreveF]  
-  
-  shocks: [ei1,ei2,ed]
-     
-  parameters:  [xi,rpi,rx,gam,pi1,pi2,pi3,mult,mult2,pir,pid,betta,i_ini,d_ini,A,theta,alfa,
-                inc_target,n_target,delta,g_ss,eta,xi_flex,pie_ss,rr_ss,Rb_ss,lockdown_policy,
-                sigma,theta_lockdown,vaccination_policy,vaccination_rate,
-                virus_resistant_strain,virus_variant_start,ex]
-
-calibration:
-   i_ini: 1.e-4          # Initial seed of infection
-   d_ini: 1.e-4          # Initial number of deaths
-   
-   betta: 0.98**(1./52)   # Weekly household discount factor
-   pid: 2.8e-3 #7*0.002/14 # Weekly probability of dying
-   pir: 7*1./14-pid #0.16 # Weekly probability of recovering
[…]
