### 2025-04-24 754d2fe
Delete snowdrop/models/COVID19/ert_model.yaml.txt
