+++ snowdrop/models/COVID19/ert_model.yaml.txt
@@ -1,462 +1,0 @@
-name: Eichenbaum, Rebelo and Trabandt model.  Application to forecasting Covid-19 economic impact.
-
-# This file solves and simulates the model developed by 
-# Eichenbaum M.S., Rebelo S., Trabandt M. (2020) “Epidemics in the Neoclassical and New Keynesian Models”.
-# NBER Working Paper 27430, http://www.nber.org/papers/w27430
-
-# Based on parameters, the code below nests the models with 
-# perfect competition (gam -> 1, xi = 0), imperfect competition with 
-# flexible prices (gam = 1.35, xi = 0) and imperfect competition with 
-# sticky prices (gam = 1.35, xi = 0.98 (weekly calibration)).
-# gam and xi are set in lines 56 and 60.
-
-# We solve the fully nonlinear model. Given the strong nonlinearities of the 
-# model due to the epidemic dynamics we use homotopy for a number of 
-# parameters, i.e. we increase parameter values stepwise until we impose
-# their final desired value. Homotopy calculations can take a while. 
-
-
-symbols:
-
-  variables: [
-    # Endogenous variables of the actual (sticky price) economy
-    y,k,n,w,rk,x,c,s,i,r,ns,ni,nr,cs,ci,cr,tau,
-    lambtilde,lamtau,lami,lams,lamr,dd,pop,Rb,pie,mc,F,Kf,rr,pbreve,v,
-    # Endogenous variables of the flexible price economy
-    yF,kF,nF,wF,rkF,xF,cF,sF,iF,rF,nsF,niF,nrF,csF,ciF,crF,
-    tauF,lambtildeF,lamtauF,lamiF,lamsF,lamrF,ddF,popF,RbF,pieF,
-    mcF,FF,KfF,rrF,pbreveF,vF]  
-
-  shocks: [ei,ed,shk_y,shk_yF]
-     
-  parameters:  [xi,rpi,rx,gam,pi1,pi2,pi3,pir,pid,betta,i_ini,d_ini,A,theta,alfa,
-                inc_target,n_target,delta,g_ss,eta,xi_flex,pie_ss,rr_ss,Rb_ss,
-                lockdown_policy,theta_lockdown,vaccination_policy,vaccination_rate,
-                virus_resistant_strain,virus_variant_start]
-
-calibration:
-   i_ini: 2.e-3          # Initial seed of infection
-   d_ini: 3.e-4          # Initial number of deaths
-   
-   betta: 0.98**(1./52)   # Weekly household discount factor
-   pid: 2.5e-3 #7*0.002/14 # Weekly probability of dying
-   pir: 7*1./14-pid #0.16 # Weekly probability of recovering
-   delta: 0.06/52        # Capital depreciation rate (weekly)
-   alfa: 2./3            # Labor share
-
-   # final numbers for pi1,pi2,pi3 and xi will be imposed below using homotopy
-   pi1: 2.e-7
-   pi2: 2.e-4
-   pi3: 0.5
-   pi1_final: pi1 
-   pi2_final: pi2 
-   pi3_final: pi3 
-   xi_final: xi 
-   
[…]
