### 2025-04-24 98e7535
Delete snowdrop/models/TOY/RBC.yaml.txt
