### 2025-04-25 cc10b40
Delete snowdrop/docs/README.md
