### 2025-04-25 a62a578
Update info.rst
