+++ README.md
@@ -12,8 +12,8 @@ This package is available at PyPi repository and can be installed via pip instal
 pip install pysnowdrop --upgrade
  
  ## How to run:
- - Create or modify existing YAML model file in snowdrop/models folder.
- - Open snowdrop/src/tests/test_toy_models.py file and set *fname* to the name of this model file.
+ - Create or modify existing YAML model file in *snowdrop/models* folder.
+ - Open *snowdrop/src/tests/test_toy_models.py* file and set *fname* to the name of this model file.
  - Run simulations in Spyder IDE by double-clicking on run button or run python script in a command prompt.
 
 ## Content:
