+++ CONTRIBUTING.md
@@ -1,0 +1,20 @@
+# Contributing
+*Snowdrop* is an open source project and welcomes contributions! Here are some ways to help:
+
+* Writing and improving the documentation
+* Reporting or fixing bugs
+* Contributing features or enhancements
+* Creating *Snowdrop* extensions for more domains
+* Creating demonstration projects or tutorials
+
+
+# Reporting Bugs
+Bugs and issues can be reported in the issue tracker using the [bug report form](https://github.com/gumilevskij/framework/new). Please check the bug has not already been reported by [searching the issue tracker](https://github.com/gumilevskij/framework/issues) before submitting a new issue!
+
+
+# Feature Requests
+Feature requests can be made in the issue tracker using the [feature report form](https://github.com/gumilevskij/framework/issues/new). 
+
+
+# Solving Issues
+To submit bug fixes, open a new GitHub pull request with the bug fix. The PR description should clearly describe the problem and solution and should include the relevant issue numbers if applicable. Instructions for how to set up a local *Snowdrop* development environment and run the tests can be found in the [developers documentation](https://gumilevskij/framework/snowdrop/docs/UserGuide.pdf).
