+++ joss_submission/paper.bib
@@ -25,6 +25,7 @@
   author = {Botman, Dennis PJ and Karam, Philippe D and Laxton, Douglas and Rose, David},
   year = {2007},
   publisher = {IMF Working Paper},
+  doi = {10.5089/9781451867640.001},
   url = {https://www.imf.org/en/Publications/WP/Issues/2016/12/31/DSGE-Modeling-at-the-Fund-Applications-and-Further-Developments-21265}
 }
 
@@ -46,6 +47,7 @@
   journal = {FRB of New York Staff Report},
   number = {647},
   year = {2013},
+  doi = {10.2139/ssrn.2344312},
   url = {https://www.newyorkfed.org/medialibrary/media/research/staff_reports/sr647.pdf}
 }
 
@@ -53,7 +55,7 @@
   title = {DSGE models used by policymakers: A survey},
   author = {Yagihashi, Takeshi},
   journal = {Policy Research Institute, Ministry of Finance, Japan, PRI Discussion Paper Series},
-  year={2020},
+  year = {2020},
   doi = {10.3390/econometrics4010001},
   url = {https://www.mof.go.jp/pri/research/discussion_paper/ron333.pdf}
 }
@@ -62,7 +64,7 @@
   author = {Michel Juillard and Douglas Laxton and Peter McAdam and Hope Pioro},
   journal = {Journal of Economic Dynamics and Control, No. 22, pp.1291—1318},
   title = {An algorithm competition: First-order iterations versus Newton-based technique},
-  year = {1998},
+  year = 1998,
   doi = {10.1016/S0165-1889(98)00013-X},
   url = {https://www.sciencedirect.com/science/article/abs/pii/S016518899800013X}
 }
@@ -71,7 +73,7 @@
   author = {Alexei Goumilevski and Aquiles Farias and Kadir Tanyeri},
   journal = {Proceedings of the 27th International Conference on Computing in Economics and Finance},
   title = {A New Macroeconomic Modeling Framework Applied to Assess Effects of COVID-19},
-  year = {2021},
+  year = 2021,
   url = {https://www.academia.edu/117687479/A_New_Macroeconomic_Modeling_Platform_Applied_to_Assess_Effects_of_COVID_19}
 }
 
@@ -79,6 +81,7 @@
   author = {Peter Ireland},
   journal = {NBER Working Paper, No. 10309},
   title = {Technology Shocks in the New Keynesian Model},
-  year = {2004},
+  year = 2004,
+  doi = {10.3386/w10309},
   url = {https://www.nber.org/system/files/working_papers/w10309/w10309.pdf}
 }
