+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "pysnowdrop"
-version = "1.0.3"
+version = "1.0.5"
 description = "Python Platform for DSGE Modeling"
 authors = ["Alexei Goumilevski <agoumilevski@hotmail.com>",
            "James Otterson <jjotterson@gmail.com>"]
