+++ README.md
@@ -1,7 +1,8 @@
 # Python Framework for DSGE Models
  
 ## Authors: Alexei Goumilevski and James Otterson
- 
+ ![Uploading IMG_20250509_000222210_PORTRAIT.jpg…]()
+
 ## What it is:
 This Framework aims to help economists to ease development and run of Dynamic Stochastic General Equilibrium (DSGE) models in Python environment.
 DSGE models utilize macroeconomic theory to analyze behavior in time of macroeconomic agents under uncertainty. They are used for policy analisys, forecasting,
