### 2025-05-13 #8197 @editorialbot
[REVIEW]: Snowdrop: Python Package for DSGE Modeling
**Submitting author:** @agoumilevski  (<a href="http://orcid.org/0009-0004-5574-854X">Alexei Goumilevski </a>)
**Repository:** https://github.com/gumilevskij/Framework
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.0
