### 2025-06-11 c51fa08
Update function.py
Removed unused code.
