+++ snowdrop/src/preprocessor/function.py
@@ -1,124 +1,4 @@
 import numpy as np
-
-# def get_compiled_function_and_jacobian(f_dynamic,params,y,shock,t=0,order=1,debug=False):
-#     """
-#     Returns function and Jacobian.
-    
-#     Parameters:
-#         :param f_dynamic: Model object.
-#         :type f_dynamic: Model.
-#         :param params: array of parameters.
-#         :type params: numpy.array.
-#         :param y: Array of values of endogenous variables.
-#         :type y: numpy.array
-#         :param shock: The values of shocks.
-#         :type shock: numpy.array
-#         :param t: Time index.
-#         :type t: int
-#         :param order: Order of partial derivatives of the system of equations.
-#         :type order: int
-#         :returns: Function , Jacobian, Hessian, and third order derivatives matrices.
-#     """
-#     # Build the approximation point to take the derivatives
-#     yy = np.concatenate([y[2],y[1],y[0],shock])
-    
-#     if debug:
-#         from snowdrop.src.preprocessor.f_dynamic import f_dynamic
-#         # m = dict(zip(model.symbols["variables"],y2))
-#         # print(m)
-#         # print()
-#         derivatives = f_dynamic(yy,params,order=order)
-#     else:
-#         derivatives = f_dynamic(yy,params,order=order)
-        
-#     return derivatives
-       
-
-def correct(n,max_lead_shock,min_lag_shock,indexEndogVariables,derivative):
-    """
-    Set elements of not swapped array to zero.
-
-    Parameters
-    ----------
-    n : int
-        number of endogenous variables.
-    max_lead_shock : int
-        Minimum lag shock number.
-    min_lag_shock : int
-        Mimum lead shock number.
-    indexEndogVariables : numpy ndarray
-        Index Of EndogVariables.
-    derivative : numpy ndarray
-        partial derivatives array.
-
-    Returns
-    -------
-    Corrected derivatives array.
-
-    """                                  
-    if max_lead_shock == 0 and min_lag_shock == 0:  
-        derivative[np.array(indexEndogVariables)-n] = 0
-        derivative[np.array(indexEndogVariables)+2*n] = 0
-    elif max_lead_shock < 0 and min_lag_shock == 0:  
-        derivative[np.array(indexEndogVariables)] = 0
-    elif max_lead_shock == 0 and min_lag_shock > 0:  
-        derivative[np.array(indexEndogVariables)+2*n] = 0
-        
-    return derivative
-            
-            
-def flip(n,order,indEndog,indExog,max_lead_shock,min_lag_shock,derivatives):
-    """
-    Swap indices of endogenous and exogenous variables.
-
-    Parameters:
-        :param order: Order of partial derivatives of the system of equations.
-        :type order: int
-        :param endogVariables: list
-        :type endogVariables: Endogenous variables to be flipped
-        :param exogVariables: list
-        :type exogVariables: Exogenous variables to be flipped
-        :returns: Function, Jacobian, Hessian, and third order derivatives matrices.
-    """
-    #TODO: Need testing
-    if indEndog is None or indExog is None:
-        return derivatives
-    
-    elif not len(indEndog) == len(indExog):
-        from snowdrop.src.misc.termcolor import cprint
-        cprint(f"Number of flipped endogenous varaibles {len(indEndog)} and exogenous variables {len(indExog)} is different.","red")
-        return derivatives
-    
[…]
