### 2025-06-11 940ef6c
Update nonlinear_solver.py
Corrected find_steady_state() method for flipped variables and parameters.
