### 2025-06-21 gumilevskij/Framework#1 @phumthep
[JOSS] Comments on the manuscript - Phumthep
Here are some of my thoughts on the JOSS manuscript.

1. The Summary section appears to very well describe Snowdrop to those within the field of economics. This section depicts the underlying modeling frameworks such as New Keynesian models, real business cycle models, etc. As JOSS welcomes submissions from broad research areas, the summary section could benefit from taking a step back and discuss the broader context of a DSGE while highlighting the functionalities of Snowdrop to the lay audience. The latter part could include expanding on Snowdrop’s robustness and versatility, and its complete suite of modeling capabilities comparable to existing DSGE models. This might benefit from moving some contents from the Statement of Need section to the summary section.I think the following points could be considered by the authors when first introducing Snowdrop in the Summary section:
- Briefly describe or define a Dynamic Stochastic General Equilibrium model
- Briefly describe the applications of DSGE models or problems these models solve – estimate the effects of economic shocks on macro-economic indicators like GDP and inflation over time
- Expand on robust and efficient solution techniques
- The implications of the following terms remain unclear to a lay audience: “robust and versatile”. It is also useful to differentiate the differences between “estimation, simulation, and forecasting”
- Move highly technical terms to the later sections
- Hopefully, the reader should understand the following point when reading the summary section: “Given [inputs], users can leverage Snowdrop to [applications].”

2. The Statement of Needs section clearly identifies current modeling frameworks: DYNARE and IRIS while pointing out that they are either commercial software or requires one as a dependency. After reading this section, a reader grasps the message that Snowdrop is a fully featured, open-source alternative. Overall, I think this section might benefit from the followings – which are for the authors to consider:
- L24-27: Since GPL is not the sole open-source licensing option as listed [here]( https://opensource.org/licenses-old/category), would it make sense to generalize?
- L27-28: The mention of Snowdrop being written in Python repeats the Summary paragraph.
- L27-28: Instead of closing this paragraph with “is intended to frill that void”, it could be more powerful to end this paragraph with what Snowdrop enables the target user to do or benefit from.

3. The Highlights section sets out to list features of Snowdrop. I think the presentation of these features could be presented in a more organized fashion as a story in the form of paragraphs, or somehow a categorized grouping of related features. This is just an idea, but the categories could include ease-of-use, solution algorithms, analysis, etc. Ideally, this section should substantiate the claims first mentioned in the Summary section.

4. The Highlights section appears to contain some features that a user likely expects by default like Snowdrop being platform neutral. The mention of Snowdrop being a Python package that only uses open-source packages potentially leads to questions about permissiveness that comes with different open-source licensing options.

5. For software that performs complex scientific calculations, it might be worth pointing out the testing and validation of algorithms. Furthermore, a reader might be curious about the computational performance as the software is written in Python instead of a low-level programming language. These points might be integrated into the manuscript.

6. To ensure that the manuscript, which is a static document, does not become obsolete, the authors could consider hosting detailed examples like the format of input files and example code in a dedicated webpage on GitHub or other location, rather than in the manuscript itself. This would yield benefits including: (1) any future changes to input file format or functions/method calls do not impact the manuscript (2) the users will have somewhere to access up-to-date information (3) make space for describing other aspects of Snowdrop.

7. The last section, titled “Status”, is a bit confusing. This section gives two analyses using Snowdrop. Would it be clearer to rename the section?

8. The manuscript uses terms like “framework”, “model”, and “toolkit”, sometimes interchangeably. For a lay audience, the distinction between these terms can be ambiguous, making the text difficult to understand. From the manuscript, it appears that Snowdrop is a collection of modeling frameworks for a variety of case studies? In energy systems modeling, a framework is a set of equations that are applied to case studies, while a model is a specific implementation of the framework, as described [here]( https://doi.org/10.1016/j.rser.2022.112719).- L66: Should “Framework” be capitalized as in the text?

9. Other minor suggestions:- L23-24 – By “rely[ing] on commercial software to run” do you mean mathematical solvers?
- L67: The terms “(solve, simulate, calibrate or estimate)” could be ambiguous to the lay audience. I think the sentence does convey the message without these brackets, otherwise maybe the draft could expand more on these terms?
- L77: Fix “employed” to “employed
- L80: A more formal language would be “through” instead of “thru”
- Figures 1 and 3 (maybe including 2) could benefit from labeling their y-axes.
