### 2025-06-21 #8197 @phumthep
Hi. I have started reviewing the submission. Overall, the software shows substantial scholarly effort as required by JOSS. Specifically, the software contains a suite of useful features, making Snowdrop likely going to be used by other researchers to do economic analysis. Although I am not a believer that the number of commits is an indication of effort, the repository has 220+ commits, so this should be well more than sufficient.

I have completed my revision of the manuscript. My suggestions are open to discussion. Building on this point, I have created an issue on the Repo related to the manuscript [here](https://github.com/gumilevskij/Framework/issues/1#issue-3165430563).

I will now proceed with the code.
