### 2025-06-22 gumilevskij/Framework#2 @phumthep
[JOSS] Improvements for Installation Guide and Dependency Management
While I was following the installation guide, I encountered a few installation issues.

**=== Issues ===**

1. Failing to run the command `pip install -r requirements.txt`

The command produces an error:

```base
ERROR: Could not find a version that satisfies the requirement python>=3.10 (from versions: none)
ERROR: No matching distribution found for python>=3.10
```

This is because `pip` is a package installer and cannot install a Python interpreter. My approach was to remove `python>=3.10` from `requirements.txt`, which allows pip to proceed.

_Also, from my understanding, this command only installs the dependencies and does not install Snowdrop package itself._ A user will not be able to import Snowdrop in their Python interpreter.

2. Dependency conflicts when installing from `requirements.txt`

After removing the python requirement from requirements.txt and running `pip install -r requirements.txt`, the installation proceeds but produces dependency conflicts.

```bash
ERROR: pip's dependency resolver does not currently take into account all the packages that are installed. This behaviour is the source of the following dependency conflicts.
automat 20.2.0 requires attrs>=19.2.0, but you have attrs 18.2.0 which is incompatible.
jsonschema 4.23.0 requires attrs>=22.2.0, but you have attrs 18.2.0 which is incompatible.
referencing 0.30.2 requires attrs>=22.2.0, but you have attrs 18.2.0 which is incompatible.
twisted 23.10.0 requires attrs>=21.3.0, but you have attrs 18.2.0 which is incompatible.
```

3. Missing dependencies

When I tried to run `test_toy_models.py`, I encountered the following `ModuleNotFoundError` when importing the followings: `dill` and `yaml`.

**=== Recommendations ===**

1. Review `requirements.txt`
2. Actually, is there a specific reason to keep requirements.txt? In this case, we have two sources of truth. Ensuring alignment between `pyproject.toml` and `requirements.txt` requires attention to detail – unless we automatically create `requirements.txt` from `pyproject.toml`. Failure to do so could lead to headaches with this “works on my machine”
3. It would be helpful to explain the use case for each of the three installation options. It is not clear for which situation would I need to adopt a particular installation option.
