### 2025-06-22 #8197 @phumthep
I noticed that we need to manually run tests for each module while test scripts are scattered across different modules. For example, `graph` module contains `test_animation.py` and `test_contour.py`. While `src/tests` contains other test scripts. Is it possible to collect them in a single location? Also, is there a way to automate the test scripts? I noticed that `pytest` is a dependency in `pyproject.toml` but I don't how to use it.
