### 2025-06-22 #8197 @phumthep
Related to the previous comment, is it possible to restructure the repository -- moving non-code assets out from `src/` ? Example project structure can be found [here](https://packaging.python.org/en/latest/tutorials/packaging-projects/) and [here](https://critical-infrastructure-systems-lab.github.io/manual/docs/general-programming#how-to-organize-a-github-repository).

Another issue is that when a user runs pip install pysnowdrop, non-code assets (PDFs, data files for case studies, and graphs) are copied deep into their Python environment directory, making the installed package unnecessarily large. Users who just want to use the core library are forced to download documentation and example data they may likely never use. Even when the user want access to these files, they are not easily discoverable due to the files being buried deep in the environment manager’s directory. The resulting filepaths are not intuitive. Therefore, the source distribution (sdist) and wheel (wheel) should contain the code and only the essential data required for the code to run.

Here is my take on the project structure. The goal is to create a clear separation between the source code and non-code assets. Please feel free to disregard:

Snowdrop
├── pyproject.toml
├── requirements.txt
├── README.md
├── LICENSE
├── CONTRIBUTING.md
├── snowdrop/              # Source code only
│   ├── __init__.py
│   ├── driver.py
│   ├── info.py
│   ├── version.py
│   ├── epidemic/
│   ├── graphs/
│   ├── gui/
│   ├── misc/
│   ├── model/
│   ├── numeric/
│   ├── preprocessor/
│   ├── samples/
│   ├── tests/
│   └── utils/
├── test_snowdrop/     # For tests and the necessary test data
├── data/                  # Data files
├── models/             # Model files
├── docs/                 # Documentation
├── bin/                   # Executables
├── results/               # Output files
├── graphs/                # Generated graphs
├── joss_submission/
└── api_docs/

I haven't looked into the source code, but this implementation may require functions and classes to accept a user's file paths, _both inputs and outputs_, as arguments, rather than hardcoded relative paths. Test data could still be kept as part of the package.

Currently, it seems that a user cannot specify the output directory for graphs and generated pdfs?
