### 2025-06-22 #8197 @gumilevskij
Dear @phumthep. Thanks for reviewing this submission. I will try to resolve all the issues, but here is a quick one. You can change directory to the "snowdrop/src" folder and run "pytest tests" in the command line. This will find the test functions of python scripts and run all the tests in that folder.
