### 2025-06-23 #8197 @gumilevskij
Dear @phumthep.  The user can easily specify the convergence criteria with these two lines of code, for example:

from snowdrop.src.numeric.solver import nonlinear_solver as ns
ns.TOLERANCE = 1.e-7

Alternatively, the user can pass the tolerance level as a parameter to the "homotopy_solver" function.  The "epidemic/ert.py" script shows an example.
I would recommend using the "simulate" function and not calling the first or the second order solvers directly, since this function abstracts and encapsulates the details of the specific solver.
The variable "itr" indicates the number of calls to the objective function of the steady state solution. It is for internal use and should not be changed by the user.
The only problem with these global variables could be race conditions when multiple threads try to modify these convergence criteria variables.  But this is not the case, since the tolerance level or the number of iterations for a given solver is constant.

Thank you.
