### 2025-06-30 gumilevskij/Framework#4 @phumthep
[JOSS] Feedback on Snowdrop tests
I have run the tests using `pytest` and encountered a few issues. Here is the summary output from running `pytest` in the `snowdrop/src` folder:

```bash
===================== short test summary info =====================
FAILED tests\test_model_estimation.py::test - ImportError: cannot import name 'pi' from 'scipy'
FAILED tests\test_toy_models.py::test_1 - _tkinter.TclError: Can't find a usable tk.tcl...
ERROR tests\test_models_import.py::test_model_import
ERROR tests\test_models_import.py::test
================ 2 failed, 3 passed, 431 warnings, 2 errors in 79.05s ================
```

1. Tests are not passing due to import issues. Also, over 400 warnings cloud out the outputs, making it difficult to understand the log.
2. The tests require a runtime of ~80 seconds likely due to the tests generating graphs and PDFs. Ideally, the tests should run quickly. Is it necessary to generate all the graphs and PDFs for testing purposes?
3. The UserGuide (page 4) states the `tests` folder is for running "different economic models". These scripts make great case studies of using Snowdrop but not necessarily automated tests.'' Would it be possible to have a clearer separation between examples and tests? To build on this point, we could develop small tests for specific core components (the numerical methods) to ensure they work as expected. Since some modules contain the main part that appears to check the calculations, this could involve extracting the "main" part of these modules for automated testing.
