### 2025-06-30 #8197 @phumthep
I have run pytest on the existing code. Some tests are failing and producing warnings. I have created an issue on GitHub to address this point [here](https://github.com/gumilevskij/Framework/issues/4).

> Dear [@phumthep](https://github.com/phumthep). Thanks for reviewing this submission. I will try to resolve all the issues, but here is a quick one. You can change directory to the "snowdrop/src" folder and run "pytest tests" in the command line. This will find the test functions of python scripts and run all the tests in that folder.
