+++ requirements.txt
@@ -1,0 +1,38 @@
+# Mandatory dependencies
+pandas >= 2.1.4
+scipy >= 1.12.0
+numpy >= 1.26.4
+matplotlib >= 3.10.0
+sympy >= 1.12
+symengine >= 0.11.0
+emcee >= 3.1.6
+statsmodels >= 0.14.0
+pydot >= 2.0.0
+networkx >= 3.2
+particles >= 0.4
+pymcmcstat >= 1.9.1
+PyPDF2 >= 3.0.1
+corner >= 2.2.2
+seaborn >= 0.13.0
+numba >= 0.59.0
+lark >= 1.2.2
+graphviz >= 0.20.3
+filterpy >= 1.4.5
+pytest >= 7.4.0
+pykalman >= 0.9.7
+pylatex >= 1.4.2
+pdfkit >= 1.0.0
+xlsxwriter >= 3.2.0
+ruamel.yaml == 0.17.21
+dill >= 0.3.8
+PyYAML >= 6.0.1
+
+# Optional dependencies
+compecon >= 2024.5.19
+quantecon >= 0.7.2
+reportlab >= 4.2.2
+textwrap3 >= 0.9.2
+PyLaTeX >= 1.4.2
+glob2 >= 0.7
+autograd >= 1.7.0
+pdflatex >= 0.1.3
