### 2025-07-01 37af5bc
Delete snowdrop/src/gui/temp.py
