### 2025-07-01 #8197 @gumilevskij
Dear @phumthep .  I reworked this project based on your advice. I moved the "test_model_estimation.py" script to the samples folder since it is a computationally intensive task.  The "pytest" test should run successfully. The error you saw was caused by a recent change to the "scipy" package that moved the pi constant to the "scipy.constants" directory. Thanks.
