### 2025-07-01 d154f66
Updated User Guide and README.md files.
