+++ README.md
@@ -3,40 +3,41 @@
 ## Authors: Alexei Goumilevski and James Otterson
  
 ## What it is:
-This Framework aims to help economists to ease development and run 
-of Dynamic Stochastic General Equilibrium (DSGE) models in Python environment.
+This Framework is designed to assist economists in the development and execution of Dynamic Stochastic General Equilibrium (DSGE) models within a Python environment.
 
 ## Installation:
 
-User is adviced to create a virtual environmemt in Python that isolates this installation along with its packages from the system-wide Python installation and other virtual environments. There are three options to install “Snowdrop” package: 
-1. Clone this GitHub repository to your local drive. To install required packages and libraries for this project, run in a command prompt,
+Users are advised to create a virtual environment in Python to isolate this installation and its packages from the system-wide Python installation and other virtual environments. There are three options to install the “Snowdrop” package:
+
+1. Clone this GitHub repository to your local drive. To install the required packages and libraries for this project, run
+   the following command in a command prompt:
    pip install -r requirements.txt
-2. Run command: pip install snowdrop-1.0.6-py3-none-any.whl --user
+2. Run the command: pip install snowdrop-1.0.6-py3-none-any.whl --user
 3. Install *Snowdrop* via pip installer: pip install pysnowdrop --upgrade
  
  ## How to run:
- - Create or modify existing YAML model file in snowdrop/models folder.
- - Open snowdrop/src/tests/test_toy_models.py file and set *fname* to the name of this model file.
- - Run simulations in Spyder IDE by double-clicking on run button or run python script in a command prompt.
+ - Create or modify existing YAML model file in the supplements/models folder.
+ - Open tests/test_toy_models.py file and set *fname* to the name of the model file.
+ - Run simulations in the Spyder IDE by double-clicking the run button or execute the Python script in a command prompt.
 
 ## Content:
  - Sample model file (see `<supplements/models/Toy/JLMP98.yaml>`)
  - Documentation (see `<supplements/docs/UserGuide.pdf>`)
 
 ## Highlights:
-- Framework is written in Python language and uses only Python libraries that are available by installing Anaconda distribution.
-- Framework is versatile to parse model files written in a human readable YAML format, Sirius XML format and to parse simple IRIS and DYNARE model files.
-- Prototype model files are created for non-linear and linear perfect-foresight models.
-- It can be run as a batch process, in a Jupyter notebook, or in a Spyder interactive development environment (Scientific Python Development environment).
-- Framework parses the model file, checks its syntax for errors, and generates Python functions source code.  It computes the Jacobian up to the third order in a symbolic form.
-- Non-linear equations are solved by iterations by Newton's method.  Two algorithms are implemented: ABLR stacked matrices method and LBJ forward-backward substitution method.
-- Linear models are solved with Binder and Pesaran's method, Anderson and More's method and two generalized Schur's method that reproduce calculations employed in Dynare and Iris.
-- Non-linear models can be run with time dependents parameters.
-- Framework can be used to calibrate models to find model's parameters. Calibration can be run for both linear and nonlinear models.  Framework applies Bayesian approach to maximize likelihood function that incorporates prior beliefs about parameters and goodness of fit of model to the data.
-- Framework can sample model parameters by using Markov Chain Monte Carlo affine invariant ensemble sampler algorithm of Jonathan Goodman.
-- Framework uses Scientific Python Sparse package for large matrices algebra.
[…]
