### 2025-07-02 16d94e4
Changed scipy version to 1.12.0 in requirements.
