+++ joss_submission/paper.md
@@ -17,37 +17,17 @@ affiliations:
 date: 29 January 2025
 bibliography: paper.bib
 ---
-
 # Summary
 
-At its core, `Snowdrop` is a robust and versatile Python package designed for the
-analysis of macroeconomic *Dynamic Stochastic General Equilibrium* (*DSGE*) models.
-In its entirety, this package offers an extensive framework for the study of various related
-economic models, including *New Keynesian* models, *Real Business Cycle* models, *Gap*
-models, and *Overlapping Generations* models. `Snowdrop` equips
-researchers with essential tools to address the fundamental requirements of
-these models, encompassing estimation, simulation, and forecasting processes.
-In particular, the package employs robust and efficient solution techniques to
-solve both linear and nonlinear perfect foresight models based on the rational
-expectations hypothesis, which is a critical need for many *DSGE* models.
+At its core, `Snowdrop` is a robust and versatile Python package designed for the analysis of macroeconomic *Dynamic Stochastic General Equilibrium* (*DSGE*) models. This package provides an extensive framework for studying various related economic models, including *New Keynesian* models, *Real Business Cycle* models, *Gap* models, and *Overlapping Generations* models. `Snowdrop` equips researchers with essential tools to address the fundamental requirements of these models, encompassing estimation, simulation, and forecasting processes.  In particular, the package employs robust and efficient solution techniques to solve both linear and nonlinear perfect foresight models based on the rational expectations hypothesis, which is critical for many DSGE models.
 
 # Statement of need
 
-*DSGE* models are a mainstay class of models employed by Central Banks around the
-world, informing key country monetary policy decisions [@botman], [@smets],
-[@del], [@yagihashi].  These models capture the dynamic evolution of economic variables influenced 
-by agents who respond to anticipated future outcomes in the present, necessitating the combined 
-use of specialized techniques that are not readily availabel even in the extensive list
-of Python's scientific modeling packages [@fernandez]. Currently, the two
-primary DSGE modeling toolboxes, [DYNARE](https://www.dynare.org) and
-[IRIS](https://iris.igpmn.org) are comprehensive toolsets that offer
-an user-friendly infrastructure with support to all stages of model development.
-These, and similar, applications, however, are either commercial, or rely on
-commercial software to run, and hence require expensive licensing costs. There
-is no integrated software package to our knowledge that is both flexible to
-handle a wide class of models with all required software to run the models
-available for free under the *GNU General Public Licensing* agreements. This
-Framework, built entirely on Python, is intended to fill that void.
+*DSGE* models are a fundamental class of models utilized by central banks worldwide, informing key monetary policy decisions [@botman], [@smets], [@del], [@yagihashi]. These models capture the dynamic evolution of economic variables influenced by agents who respond to anticipated future outcomes in the present. This necessitates the use of specialized techniques that are not readily available even in the extensive list of Python's scientific modeling packages [@fernandez]. The three primary DSGE modeling toolboxes currently include [DYNARE](https://www.dynare.org), [IRIS](https://iris.igpmn.org) and [TROLL](https://www.intex.com/troll/history.html). While *Dynare* and *IRIS* are free and open-source software, they were primarily developed to run on the *MATLAB* platform, which is commercial. In contrast,*TROLL*, on the other hand, is a commerial application and requires subscription.
+
[…]
