### 2025-07-02 308b61d
Modified tests script to disable graphs in pytest.
