+++ tests/test_kalman_filter.py
@@ -42,8 +42,9 @@ def test(fname='TOY/Ireland2004.yaml',fmeas='supplements/data/gpr_1948.csv'):
           
 
     ### Run Kalman filter
+    Plot = not 'pytest' in sys.modules
     model.setStartingValues(hist=meas,bTreatMissingObs=False,debug=False)
-    yy,dates,epsilonhat,etahat = kalman_filter(model=model,meas=meas,fout=fout,Output=True,Plot=True)
+    yy,dates,epsilonhat,etahat = kalman_filter(model=model,meas=meas,fout=fout,Output=True,Plot=Plot)
     
 
 if __name__ == '__main__':
