### 2025-07-02 8fa0bcf
Update pyproject.toml
