+++ pyproject.toml
@@ -43,7 +43,7 @@ Repository = "https://github.com/gumilevskij/Framework.git"
 python =  "^3.10"
 pandas =  "^2.1.4"
 scipy =  "1.12.0"
-numpy =  "^1.26.4"
+numpy =  "1.26.4"
 matplotlib =  "3.10.0"
 sympy =  "^1.12"
 symengine =  "^0.11.0"
