+++ README.md
@@ -11,13 +11,18 @@ Users are advised to create a virtual environment in Python to isolate this inst
 
 1. Clone this GitHub repository to your local drive. Then install the necessary packages and libraries for this project.
    The following commands illustrate an example:
+   
    cd to_your_working_directory
+   
    git clone https:/github.com/gumilevskij/Framework.git
+   
    python -m venv env
+   
    source env/bin/activate
+   
    pip install -r requirements.txt
-2. Run the command: pip install snowdrop-1.0.6-py3-none-any.whl --user
-3. Install *Snowdrop* via pip installer: pip install pysnowdrop --upgrade
+3. Run the command: pip install snowdrop-1.0.6-py3-none-any.whl --user
+4. Install *Snowdrop* via pip installer: pip install pysnowdrop --upgrade
  
  ## How to run:
  - Create or modify existing YAML model file in the supplements/models folder.
