+++ joss_submission/paper.md
@@ -36,7 +36,7 @@ Benchmarking the Python Framework against the *DYNARE* and *IRIS* toolboxes for 
 - `Snowdrop` models can be written in user-friendly *YAML* format, pure Python scripts, or in a combination of both.
 - Non-linear equations are solved iteratively via Newton's method. `Snowdrop` implements the *ABLR* stacked matrices and *LBJ* [@Juillard] forward-backward substitution method to solve such systems.
 - Linear models are solved with *Binder Pesaran's* method, *Anderson and More's* method and two generalized *Schur's* method that reproduce calculations employed in *Dynare* and *Iris*.
-- Non-linear models can be run with time dependents parameters.
+- Non-linear models can be run with time dependent parameters.
 - Goodness of fit of model data can be checked via the *Bayesian* approach to the maximization of likelihood functions.
 - Model parameters can be sampled via the *Markov Chain Monte Carlo* affine invariant ensemble sampler algorithm of Jonathan Goodman and an adaptive Metropolis-Hasting’s algorithms of Paul Miles. The former algorithm is useful for sampling badly scaled distributions of parameters. The later algorithm employs adaptive Metropolis methods that incorporate delayed rejection to stimulate samples’ states mixing.
 - Finally, `Snowdrop` streamlines the model production process by aiding users with the plotting and model reporting and storage process.
