+++ pyproject.toml
@@ -42,7 +42,7 @@ Repository = "https://github.com/gumilevskij/Framework.git"
 [tool.poetry.dependencies]
 python =  "^3.10"
 pandas =  "^2.1.4"
-scipy =  "1.12.0"
+scipy =  "1.11.4"
 numpy =  "1.26.4"
 matplotlib =  "3.10.0"
 sympy =  "^1.12"
