+++ tests/__init__.py
@@ -1,3 +1,2 @@
-__all__ = ["test_kalman_filter","test_model_estimation",
-           "test_optimization","test_SA_model",
-           "test_toy_models","test_ERT_model"]
+__all__ = ["test_kalman_filter","test_models_import",
+           "test_optimization","test_toy_models"]
