+++ snowdrop/src/preprocessor/f_measurement.py
@@ -20,14 +20,14 @@ def f_measurement(x, p, exog=[0], order=1, ind=None):
     _xi_1 = 0 
     _xi_2 = 0 
     _xi_3 = 0 
-    L_GDP_RW_GAP__ = x[10]
-    L_CPI__ = x[17]
-    L_GDP__ = x[24]
-    RS__ = x[30]
-    RS_RW__ = x[32]
-    DLA_CPI_RW__ = x[37]
-    D4L_CPI_TAR__ = x[45]
-    L_S__ = x[47]
+    RS__ = x[4]
+    D4L_CPI_TAR__ = x[16]
+    L_GDP__ = x[22]
+    RS_RW__ = x[23]
+    L_GDP_RW_GAP__ = x[26]
+    L_S__ = x[27]
+    L_CPI__ = x[33]
+    DLA_CPI_RW__ = x[38]
     OBS_L_GDP__ = x[49]
     OBS_L_CPI__ = x[50]
     OBS_RS__ = x[51]
@@ -63,28 +63,28 @@ def f_measurement(x, p, exog=[0], order=1, ind=None):
     
     # Jacobian: 
     jacobian = np.zeros((8,65))
-    jacobian[0,24] = -1
+    jacobian[0,22] = -1
     jacobian[0,49] = 1
     jacobian[0,57] = -1
-    jacobian[1,17] = -1
+    jacobian[1,33] = -1
     jacobian[1,50] = 1
     jacobian[1,58] = -1
-    jacobian[2,30] = -1
+    jacobian[2,4] = -1
     jacobian[2,51] = 1
     jacobian[2,59] = -1
-    jacobian[3,47] = -1
+    jacobian[3,27] = -1
     jacobian[3,52] = 1
     jacobian[3,60] = -1
-    jacobian[4,45] = -1
+    jacobian[4,16] = -1
     jacobian[4,53] = 1
     jacobian[4,61] = -1
-    jacobian[5,10] = -1
+    jacobian[5,26] = -1
     jacobian[5,54] = 1
     jacobian[5,62] = -1
-    jacobian[6,37] = -1
+    jacobian[6,38] = -1
     jacobian[6,55] = 1
     jacobian[6,63] = -1
-    jacobian[7,32] = -1
+    jacobian[7,23] = -1
     jacobian[7,56] = 1
     jacobian[7,64] = -1
 
