+++ joss_submission/paper.md
@@ -19,21 +19,20 @@ bibliography: paper.bib
 ---
 # Summary
 
-At its core, `Snowdrop` is a robust and versatile Python package designed for the analysis of macroeconomic *Dynamic Stochastic General Equilibrium* (*DSGE*) models. This package provides an extensive framework for studying various related economic models, including *New Keynesian* models, *Real Business Cycle* models, *Gap* models, and *Overlapping Generations* models. `Snowdrop` equips researchers with essential tools to address the fundamental requirements of these models, encompassing estimation, simulation, and forecasting processes.  In particular, the package employs robust and efficient solution techniques to solve both linear and nonlinear perfect foresight models based on the rational expectations hypothesis, which is critical for many DSGE models.
+At its core, `Snowdrop` is a versatile Python package designed for the analysis of macroeconomic *Dynamic Stochastic General Equilibrium* (*DSGE*) models. These *DSGE* models form a class of macroeconomic frameworks that elucidate the economy's behavior by considering the intertemporal decisions made by economic agents, including households, firms, and the government.  `Snowdrop` package provides an extensive framework for studying various related economic models, including *New Keynesian* models, *Real Business Cycle* models, *Gap* models, and *Overlapping Generations* models. `Snowdrop` equips researchers with essential tools to address the fundamental requirements of these models, encompassing estimation, simulation, and forecasting processes.  In particular, the package employs robust and efficient solution techniques to solve both linear and nonlinear perfect foresight models based on the rational expectations hypothesis, which is critical for many DSGE models.
 
 # Statement of need
 
-*DSGE* models are a fundamental class of models utilized by central banks worldwide, informing key monetary policy decisions [@botman], [@smets], [@del], [@yagihashi]. These models capture the dynamic evolution of economic variables influenced by agents who respond to anticipated future outcomes in the present. This necessitates the use of specialized techniques that are not readily available even in the extensive list of Python's scientific modeling packages [@fernandez]. The three primary DSGE modeling toolboxes currently include [DYNARE](https://www.dynare.org), [IRIS](https://iris.igpmn.org) and [TROLL](https://www.intex.com/troll/history.html). While *Dynare* and *IRIS* are free and open-source software, they were primarily developed to run on the *MATLAB* platform, which is commercial. In contrast,*TROLL*, on the other hand, is a commerial application and requires subscription.
+*DSGE* models are a fundamental class of models utilized by central banks worldwide, informing key monetary policy decisions [@botman], [@smets], [@del], [@yagihashi]. These models capture the dynamic evolution of economic variables influenced by agents who respond to anticipated future outcomes in the present.  They account for the transmission of random shocks and their effects on the economy, primarily serving to analyze the impacts of economic policy and facilitate forecasting. This necessitates the use of specialized techniques that are not readily available even in the extensive list of Python's scientific modeling packages [@fernandez]. The three primary DSGE modeling toolboxes currently include [DYNARE](https://www.dynare.org), [IRIS](https://iris.igpmn.org) and [TROLL](https://www.intex.com/troll/history.html). While *Dynare* and *IRIS* are free and open-source software, they were primarily developed to run on the *MATLAB* platform, which is commercial. In contrast,*TROLL*, on the other hand, is a commerial application and requires subscription.
 
[…]
