+++ snowdrop/src/numeric/solver/nonlinear_solver.py
@@ -82,13 +82,14 @@ def simulate(model,T,periods,y0,steady_state=None,params=None,cov=None,order=1,N
     return (count,yy,yyIter,err,elapsed)
     
  
-def second_order_approximation(model,jacobian,hessian,g_y,g_u,cov=None):
+def second_order_derivatives(model,jacobian,hessian,g_y,g_u,cov=None):
     """
     Compute the second order reduced form solution of the DSGE model.
     
-    It implements an algoritm described by Michel Juillard in:
+    It implements the algoritm described by Michel Juillard in
     "Computing first and second order approximation of DSGE models with DYNARE", CEPREMAP
     
+    For references please see: 
     https://archives.dynare.org/events/paris-1005/first-second-order.pdf%3Fmonth:int=12&year:int=2018&orig_query=
     https://stephane-adjemian.fr/dynare/slides/dsge-perturbation-method.pdf
     
@@ -107,6 +108,8 @@ def second_order_approximation(model,jacobian,hessian,g_y,g_u,cov=None):
         :type g_y: numpy.ndarray.
         :param g_u: Matrix of shocks.
         :type g_u: numpy.ndarray.
+        :param g_xx: Second derivative of of policy function.
+        :type g_xx: numpy.ndarray.
         :param sigma: Matrix of error covariances.
         :type sigma: numpy.ndarray.
         
@@ -127,59 +130,65 @@ def second_order_approximation(model,jacobian,hessian,g_y,g_u,cov=None):
     F_yp_yp = hessian[:n,:n,:n] 
     F_ym_ym = hessian[:n,2*n:3*n,2*n:3*n] 
     F_ym_u  = hessian[:n,2*n:3*n,3*n:,]
-    F_y_u   = hessian[:n,n:2*n,3*n:,]
+    F_yp_yc = hessian[:n,:n,n:2*n] 
+    F_yp_ym = hessian[:n,:n,2*n:3*n] 
+    F_yc_ym = hessian[:n,n:2*n,2*n:3*n]
+    F_y_yp  =  hessian[:n,n:2*n,:n]
     F_u_u   = hessian[:n,3*n:,3*n:]
-    #F_yp_u  = hessian[:n,:n,3*n:,]
-    #F_u_y   = hessian[:n,3*n:,n:2*n]
-    #F_y_u   = hessian[:n,2*n:3*n:,3*n:]
+    F_ym_yc = hessian[:n,2*n:3*n,n:2*n]
+    F_yc_u  = hessian[:n,n:2*n,3*n:]
+    F_yp_u  = hessian[:n,:n,3*n:]
             
     # Compute matrix g_y_y
-    A     = F_yp @ g_y + F_yc
-    B     = F_yp
-    C     = la.kron(g_y, g_y)
-    D     = F_ym_ym + F_yc_yc @ g_y @ g_y 
+    A  = F_yp @ g_y + F_yc
+    B  = F_yp
+    C  = la.kron(g_y, g_y)
+    D  = F_ym_ym 
+    D += 3*F_yc_yc @ g_y @ g_y @ g_y @ g_y 
+    D += 3*F_y_yp @ g_y @ g_y @ g_y 
+    D += F_yc_yc @ g_y 
+    D += F_yc_ym @ g_y
+    D += F_yc_yc @ g_y @ g_y
+    D += 3*F_yp_ym @ g_y @ g_y
+    D += F_yp_yc @ g_y @ g_y @ g_y 
+    
     # Solve Sylvester equation: A*x + B*x*C = D
     g_y_y = sylvester_solver(A=A,B=B,C=C,D=-D)
 
     # Compute matrix g_y_u
[…]

[…]
