### 2025-07-09 38e7449
Fixed duplcate functions in model.py and util.py
