+++ snowdrop/src/model/util.py
@@ -389,22 +389,14 @@ def getLimits(var_names,constraints,cal):
         for c in constraints:
             lb = ub = None
             if v in c:
-                if '.lt.' in c:
+                if '.le.' in c or '<=' in c:
                     Iu = True
-                    ind = c.index('.lt.')
-                    s = c[4+ind:].strip()
-                    if s in cal:
-                        val = cal[s]
+                    if '.le.' in c:
+                        ind = c.index('.le.')
+                        s = c[4+ind:].strip()
                     else:
-                        try:
-                            val = float(s)
-                        except:
-                            val = np.inf
-                    ub = float(val)-1.e-10
-                elif '.le.' in c:
-                    Iu = True
-                    ind = c.index('.le.')
-                    s = c[4+ind:].strip()
+                        ind = c.index('<=')
+                        s = c[2+ind:].strip()
                     if s in cal:
                         val = cal[s]
                     else:
@@ -413,22 +405,30 @@ def getLimits(var_names,constraints,cal):
                         except:
                             val = np.inf
                     ub = float(val)
-                elif '.gt.' in c:
-                    Il = True
-                    ind = c.index('.gt.')
-                    s = c[4+ind:].strip()
+                elif '.lt.' in c or '<' in c:
+                    Iu = True
+                    if '.lt.' in c:
+                        ind = c.index('.lt.')
+                        s = c[4+ind:].strip()
+                    else:
+                        ind = c.index('<')
+                        s = c[1+ind:].strip()
                     if s in cal:
                         val = cal[s]
                     else:
                         try:
                             val = float(s)
                         except:
-                            val = -np.inf
-                    lb = float(val)+1.e-10
-                elif '.ge.' in c:
+                            val = np.inf
+                    ub = float(val)-1.e-10
+                elif '.ge.' in c or '>=' in c:
                     Il = True
-                    ind = c.index('.ge.')
-                    s = c[4+ind:].strip()
+                    if '.ge.' in c:
+                        ind = c.index('.ge.')
+                        s = c[4+ind:].strip()
+                    else:
+                        ind = c.index('>=')
+                        s = c[2+ind:].strip()
                     if s in cal:
                         val = cal[s]
                     else:
@@ -437,9 +437,29 @@ def getLimits(var_names,constraints,cal):
                         except:
                             val = -np.inf
                     lb = float(val)
-                elif '.eq.' in c:
-                    ind = c.index('.eq.')
-                    s = c[1+ind:].strip()
+                elif '.gt.' in c or '>' in c:
+                    Il = True
+                    if '.gt.' in c:
+                        ind = c.index('.gt.')
+                        s = c[4+ind:].strip()
+                    else:
+                        ind = c.index('>')
+                        s = c[1+ind:].strip()
+                    if s in cal:
+                        val = cal[s]
+                    else:
+                        try:
+                            val = float(s)
+                        except:
+                            val = -np.inf
+                    lb = float(val)+1.e-10
+                elif '.eq.' in c or '=' in c:
+                    if '.eq.' in c:
+                        ind = c.index('.eq.')
+                        s = c[4+ind:].strip()
+                    else:
+                        ind = c.index('=')
+                        s = c[1+ind:].strip()
                     if s in cal:
                         val = cal[s]
                         lb = ub = val
@@ -448,6 +468,7 @@ def getLimits(var_names,constraints,cal):
                             val = float(s)
                         except:
[…]
