+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "snowdrop"
-version = "1.0.6"
+version = "1.0.7"
 description = "Python Platform for DSGE Modeling"
 authors = ["Alexei Goumilevski <agoumilevski@hotmail.com>",
            "James Otterson <jjotterson@gmail.com>"]
