### 2025-07-11 97b7e8e
Added splcge model.
