### 2025-07-11 a980a3a
Added stdcge.yaml model file.
