+++ snowdrop/src/numeric/optimization/util.py
@@ -375,9 +375,7 @@ def getIndex(e,ind):
        
     return index
     
-
-
-
+    
 def expand_map(sets,indices,m):
     """
     Iterates thru a list of indices and categories, 
@@ -400,9 +398,9 @@ def expand_map(sets,indices,m):
     # Loop over indices
     for i,index in enumerate(indices):
         for k in m:
+            values = m[k]
             sub = "("+index+")"
-            if sub in k:
-                values = m[k]
+            if sub in k or sub in str(values):
                 for j,c in enumerate(sets[index]):
                     key = replace_all(sub,"_"+c,k)
                     if isinstance(values,list) and j < len(values):
@@ -433,11 +431,11 @@ def expand_map(sets,indices,m):
                             op = values[ind1+5:ind2-1] 
                             txt = expand_prod(sets,indices,op) 
                             values = values[:ind1] + txt + values[ind2:] 
-                        if not key in out: 
-                            out[key] = replace_all(sub,"_"+c,values)
+                        out[key] = replace_all(sub,"_"+c,values)
                     else:
                         if not key in out: 
                             out[key] = values 
+                        
             else:
                 if not k in out: 
                     out[k] = m[k]
@@ -521,204 +519,6 @@ def getLabels(keys,m):
             labels[k] = m[k]
             
     return labels    
-    
-        
-def importModel(fpath):
-    """
-    Parse a model file and create a model object.
-    
-    Parameters:
-        :param fpath: Path to model file.
-        :type fpath: str.
-        
-    """
-    global eqs_labels
-    import re
-    from model.symbolic import SymbolicModel
-    from model.model import Model
-    
-    name = "Model"
-    solver = None; method = None
-    symbols = {}; calibration = {}; constraints = {}; obj = {}; labels = {}; options = {}
-    variables = []; parameters = []; equations = []
-    
-    with open(fpath,  encoding='utf8') as f:
-        txt = f.read()
-        txt = txt.replace('^', '**')
-        data = yaml.load(txt, Loader=yaml.Loader)
-        # Model name
-        name = data.get('name','GAMS model')
-        # Model equations to solve
-        model_eqs = data.get('Model',[])
-        # Solver
-        solver = data.get('Solver',None)
-        # Method
-        method = data.get('Method',None)
-        # Sets section
-        _sets = data.get('sets',{})
-        indices = [x.split(" ")[-1].split("(")[0].strip() for x in _sets.keys()]
-        sets = {}
-        for k in _sets:
-            arr = list(filter(None,k.split(" ")))
-            k1 = k[:-len(arr[-1])].strip()
-            indx = arr[-1].strip()
-            if "(" in indx and ")" in indx:
-                ind1 = indx.index("(")
-                ind2 = indx.index(")")
-                k2 = indx[1+ind1:ind2].strip()
-                k3 = indx[:ind1].strip()
-            else:
-                k2 = None
-                k3 = indx
-            if isinstance(_sets[k],str) and _sets[k] in sets:
-                sets[k3] = sets[_sets[k]]
-            else:
-                sets[k3] = _sets[k]
-            # Check that all elements of map for key=k3 are subset of elements of this map for key=k2
-            if not k2 is None:
-                diff = set(sets[k3]) - set(sets[k2])
-                if len(diff) > 0:
-                    diff = ",".join(diff)
-                    cprint(f"\nMisspecified elements of set '{k1}': extra elements - {diff}.","red")
-                    sys.exit()
-                
-        # Symbols section
-        symbols = data.get('symbols',{})
-        variables = symbols.get('variables',[])
-        parameters = symbols.get('parameters',[])
-        
-        # Equations section
-        eqs = data.get('equations',[])
-        equations,complementarity = fix(eqs,model_eqs)
-        if not len(eqs) == len(equations):
[…]
