+++ supplements/models/OPT/armington.yaml
@@ -1,4 +1,7 @@
 name: Armington Trade Equilibrium with Iceberg Costs
+# Ch.23. Computing General Equilibrium Theories of Monopolistic Competition and Heterogeneous Firms
+# by Edward J. Balistreri, Thomas F. Rutherford
+# https://www.sciencedirect.com/science/article/pii/B9780444595683000237
 
 sets:
    regions r:   [R1, R2, R3]
@@ -45,6 +48,7 @@ calibration:
     Y(G1)(R1)     : 3.5
     Y(G1)(R2)     : 3.0
     tau(j)(r)(s)  : 1.5
+    zeta(j)(r)(s) : 1
       
 constraints:
    # Positive Variables
@@ -82,6 +86,6 @@ labels:
     SUP:  Input supply (output)
     
 Model: [DEM.Q,ARM.P,MKT.c,SUP.Y]
-Solver: 'CONSTRAINED_OPTIMIZATION'   # 'MCP', 'CONSTRAINED_OPTIMIZATION', 'ROOT'
+Solver: 'ROOT'   # 'MCP','PATH', 'CONSTRAINED_OPTIMIZATION', 'ROOT'
 
     
