### 2025-07-12 #8197 @phumthep
I have closed [the issue](https://github.com/gumilevskij/Framework/issues/2) related to unable to install using `pip install -r requirements.txt`.
