### 2025-07-12 gumilevskij/Framework#4 @phumthep
I have run `pytest tests` and the five tests are passing but with a significant number of warnings (1000+). At this stage, I think the code works and future tests can be improved/added later.
