### 2025-07-12 #8197 @phumthep
> Dear [@phumthep](https://github.com/phumthep). To see API methods documentation please refer to supplements/api_docs/_build/HTML/index.html file. This documentation was built with Sphinx document generator. Thanks.

Thank you! Would you be able to add this to the "Content" section of the GitHub repo? This would help with the "Functionality documentation" box.
