### 2025-07-16 e3e5eb1
Added DICE model files.
