+++ snowdrop/src/climate/DICE/dice.py
@@ -1,0 +1,298 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+The Dynamic Integrated Climate-Economy (DICE) model by William Nordhaus. 
+
+This is the beta version of DICE-2016R. The major changes are outlined in Nordhaus,
+"Revisiting the social cost of carbon: Estimates from the DICE-2016R model."
+
+Translation from Python version: https://github.com/hazem2410/PyDICE
+
+Created on July 3, 2023.
+@author: A.Goumilevski
+"""
+
+
+import os,sys
+import numpy as np
+from scipy.optimize import minimize
+#from scipy.optimize import Bounds 
+
+path = os.path.dirname(os.path.abspath(__file__))
+working_dir = os.path.abspath(path + "/../../../..")
+fdir = os.path.join(working_dir,"graphs")
+
+sys.path.append(working_dir)
+os.chdir(working_dir)
+
+from snowdrop.src.graphs.util import plot
+from snowdrop.src.driver import run,importModel
+
+it = 0; model = None
+
+def runDICEmodel(file_path,data_path):
+    
+    output_variables = ["E","FORC","C","DAMAGES","DAMFRAC","ABATECOST","Y","TATM","TOCEAN"]       # List of variables for which decomposition plots are produced
+    decomp_variables = ["Y","DAMFRAC","ABATECOST","E"]
+
+    # Create model object
+    model = importModel(fname=file_path,use_cache=False)
+
+    T = model.options["T"]
+    shocks = model.symbols["shocks"]
+    n_shk = len(shocks)
+    variables_names  = model.symbols["variables"]
+    ind_CEMUTOTPER = variables_names.index("CEMUTOTPER")
+    variables_values = model.calibration["variables"].copy()
+    variables = dict(zip(variables_names,variables_values))
+    param_names = model.symbols["parameters"]
+    param_values = model.calibration["parameters"].copy()
+    params = dict(zip(param_names,param_values))
+    var_labels = model.symbols.get("variables_labels",{})
+    
+    # Parameters
+    tstep = params["tstep"]
+    scale1 = params["scale1"]
+    scale2 = params["scale2"]
+    fex0 = params["fex0"]
+    fex1 = params["fex1"]
+    pback = params["pback"] 
+    gback = params["gback"]
+    prstp = params["prstp"]
+    dela = params["dela"]
+    gsig = params["gsig"]
+    dsig = params["dsig"]
+    Al = params["Al"]
+    sigma = params["sigma"]
+    cost1 = params["cost1"]
+    eland0 = params["eland0"]
+    deland = params["deland"]
+    expcost2 = params["expcost2"]
+    cumetree = params["cumetree"]
+    popasym = params["popasym"] 
+    popadj = params["popadj"]
+    ga0 = params["ga0"] 
+    etree = params["etree"]
+    L = params["L"]
+    optlrsav = params["optlrsav"]
+    dk = params["dk"]
+    gama = params["gama"]
+    a1 = params["a1"]
+    a2 = params["a2"]
+    a3 = params["a3"]
+    c1 = params["c1"]
+    c3 = params["c3"]
+    c4 = params["c4"]
+    b11 = params["b11"]
[…]
