+++ snowdrop/src/numeric/optimization/optimize.py
@@ -250,7 +250,6 @@ def solver(model):
             # if solver == 'PATH':
             #     x = x0
             #     f = np.zeros(n) 
-            #     os.environ['PATH_LICENSE_STRING'] = '2830898829&Courtesy&&&USR&45321&5_1_2021&1000&PATH&GEN&31_12_2025&0_0_0&6000&0_0'
             #     # Load path library
             #     libc = loadLibrary()
             #     # Call path solver C++ function
