### 2025-07-17 d6e528e
Updated manuscript.
