+++ joss_submission/paper.md
@@ -27,7 +27,7 @@ At its core, `Snowdrop` is a versatile Python package designed for the analysis 
 
 Each of these applications has its own advantage. The ease of use through a user-friendly interface, combined with the capability to handle a variety of models, has led to the immense popularity of *DYNARE* among general equilibrium modelers. However, *DYNARE* can only handle stationary *DSGE* models and requires users to write models in a stationary format by introducing variable deflators. The *IRIS* macroeconomic toolbox is another excellent tool that has gained popularity among economists for analyzing non-stationary *DSGE* models. *TROLL*, on the other hand, specializes in efficiently solving and simulating large systems of equations. All these applications, however, are either commercial or are relying on commercial software that requires expensive licensing costs. To our knowledge, there is no integrated software package that is flexible enough to handle a wide range of models and available for free under the GNU General Public License agreements. This framework aims to fill that void. 
 
-`Snowdrop` is designed to assist economists and others in developing and analyzing complex economic problems.   Benchmarking this software against the *DYNARE* and *IRIS* toolboxes for small to medium-sized models with several hundred equations demonstrates comparable CPU execution times but a smaller memory footprint, due to the significant memory requirements of *MATLAB*.
+`Snowdrop` is designed to assist economists and others in developing and analyzing complex economic problems.   Benchmarking this software against the *DYNARE* and *IRIS* toolboxes for small to medium-sized models with several hundred equations demonstrates comparable CPU execution times but a smaller memory footprint, due to the significant memory requirements of *MATLAB*. `Snowdrop` implements several methods to solve model equations that have been tested against solutions of *DYNARE* and *IRIS*.
 
 
 ## Highlights
@@ -80,7 +80,7 @@ This toolkit provides users with an integrated Framework to input their models, 
 
 ![Monetary Policy Example\label{fig:1}](Decomposition.png)
 
-Another example illustrates the economic effects of the pandemic. We used the Eichenbaum-Rebelo-Trabandt (ERT) model [@Eichenbaum], which embeds epidemiological concepts into a New Keynesian modeling framework. We assumed that there are two strains of pathogens and employed a Suspected-Infected-Recovered (SIR) epidemiological model. These epidemiological equations were incorporated into the ERT model, consisting of sixty-four equations of macroeconomic variables from sticky and flexible price economies. The macroeconomic variables of these two economies were linked through a Taylor rule equation for the policy interest rate. The model is highly non-linear and is solved using a homotopy method, where parameters are adjusted step-by-step. We assumed that government containment measures were more lenient during the second strain of the virus compared to the first one. Figures 2 and 3 illustrate the forecast of virus transmission and deviations of macroeconomic variables from their steady state.
+Another example illustrates the economic effects of the pandemic. We used the Eichenbaum-Rebelo-Trabandt (ERT) model [@Eichenbaum], which embeds epidemiological concepts into a New Keynesian modeling framework. We assumed that there are two strains of pathogens and employed a Suspected-Infected-Recovered (SIR) epidemiological model. These epidemiological equations were incorporated into the ERT model, consisting of sixty-four equations of macroeconomic variables for sticky and flexible price economies. The macroeconomic variables of these two economies were linked through a Taylor rule equation for the bond interest rate. The model is highly non-linear and is solved using a homotopy method, where parameters are adjusted step-by-step. We assumed that government containment measures were more lenient during the second strain of the virus compared to the first one. Figures 2 and 3 illustrate the forecast of virus transmission and deviations of macroeconomic variables from their steady state.
 
 ![Epidemic Forecast\label{fig:2}](Virus.png)
 
