### 2025-07-17 gumilevskij/Framework#1 @agoumilevski
Hi @phumthep.  I tried to address your comments regarding this manuscript.  I expanded the section on DSGE class of models.  I use framework term and removed platform and model ones. I renamed Status section to Results. I also did numerous other corrections. This package results were tested against Dynare and Iris.  The test_toy_models.py script displays model equations residuals when using the numerical solution.  These residuals are of the order of 1e-20 - 1e-18.  This is an accuracy of model solution.  Thanks.
