### 2025-07-17 4e31bf8
Minor changes to paper.md file.
