### 2025-07-26 #8197 @phumthep
Thank you for the wait. I've been traveling for the past week. I've ticked all boxes on my checklist.

Nevertheless, I think the installation guide remains unclear but can be revised at a later time. Specifically, a potential user would likely still have questions about the different approaches to install the package after reading the following description:

> The first method enables access to the source code and easy update via the git pull command. The second method provides an easy installation of this package and simple update.

Both seems to provide an "easy" installation and an update. It would be great to differentiate these approaches.
