### 2025-07-26 #8197 @gumilevskij
Thank you, @phumthep.  I modified installation directives to differentiate three methods.
Dear @dazhwu, I added presentation slides explaining modeling approaches and comparison with
the alternative software.  Please let me know if I missed anything.
