### 2025-07-27 #8197 @dazhwu
I believe the current version of the paper is ready to be published.
