### 2025-07-31 5510cec
update installation and quick usage
