### 2025-07-31 #8197 @Nikoleta-v3
👋🏻  @agoumilevski @gumilevskij

I think I might have asked this before, but is there a particular reason the repository is called Framework and not `pysnowdrop` or `snowdrop`? It seems a bit inconsistent with the pip package name, and I’m curious if there’s a reason behind that.

I also agree with @phumthep  that the installation instructions, and other parts of the README, could use some improvement.

For example, the current instructions mention that “the third method requires downloading a wheel file from this repository,” but it’s not clear where in the repo that file actually is. As it stands, the command

```
pip install snowdrop-1.0.7-py3-none-any.whl --user
```
is broken unless the user knows where to find the file and navigates to the folder.

More broadly, the README assumes the user has already cloned the repository. It doesn't provide any standalone guidance for using the package in other contexts. In the paper, you mention that model specifications can be written in YAML, Python scripts, or a combination of both, which is great, but those details are buried in the PDF user guide inside the docs folder.

I’d suggest renaming the “How to run” section to something like Quick Start, since it’s not a full guide on how to use the package—more like a fast way to clone and run the toy models. There's nothing there about creating your own YAML model or using the Python API directly, even though those options are discussed in the documentation (see previous comment).

To help out, I’ve opened a pull request with a few suggested improvements to the README: https://github.com/gumilevskij/Framework/pull/5. Feel free to merge it as-is or modify it if you’d prefer a different approach.

On the topic of documentation, I’d strongly recommend moving away from `.pdf` or `.docx` formats and toward something more standard and accessible, like Sphinx based docs or even GitHub’s built-in Wiki. PDF and Word documents are harder to browse, link to, or search, especially for users working online. Anyway this is a recommendation.

A few other small things:

- In your bibliography, some references aren’t compiling properly due to capitalization issues. You’ll need to use curly braces ({}) in BibTeX to preserve case.

- There’s a small typo in the manuscript:
“the later algorithm employs adaptive Metropolis methods” → should be “the latter algorithm…”

Once you’ve addressed these points and made any final adjustments, I’d suggest publishing a new release of the software, especially if the changes since reviewing started were significant. If you do, please create a new Zenodo archive as well. When that's ready, just let me know which version to cite, along with the DOI and Zenodo link.

Thanks!
