### 2025-07-31 b48b9c5
add a few sentences to the documentation part
