+++ supplements/api_docs/source/src.misc.rst
@@ -1,61 +1,0 @@
-src.misc package
-================
-
-Submodules
-----------
-
-src.misc.display module
------------------------
-
-.. automodule:: src.misc.display
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-src.misc.info module
---------------------
-
-.. automodule:: src.misc.info
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-src.misc.latex module
----------------------
-
-.. automodule:: src.misc.latex
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-src.misc.linter module
-----------------------
-
-.. automodule:: src.misc.linter
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-src.misc.termcolor module
--------------------------
-
-.. automodule:: src.misc.termcolor
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-src.misc.text2latex module
---------------------------
-
-.. automodule:: src.misc.text2latex
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: src.misc
-   :members:
-   :undoc-members:
-   :show-inheritance:
