+++ supplements/api_docs/source/src.numeric.solver.rst
@@ -1,109 +1,0 @@
-src.numeric.solver package
-==========================
-
-Submodules
-----------
-
-src.numeric.solver.ABLR module
-------------------------------
-
-.. automodule:: src.numeric.solver.ABLR
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-src.numeric.solver.AndersonMoore module
----------------------------------------
-
-.. automodule:: src.numeric.solver.AndersonMoore
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-src.numeric.solver.Benes module
--------------------------------
-
-.. automodule:: src.numeric.solver.Benes
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-src.numeric.solver.BinderPesaran module
----------------------------------------
-
-.. automodule:: src.numeric.solver.BinderPesaran
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-src.numeric.solver.LBJ module
------------------------------
-
-.. automodule:: src.numeric.solver.LBJ
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-src.numeric.solver.Villemot module
-----------------------------------
-
-.. automodule:: src.numeric.solver.Villemot
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-src.numeric.solver.linear\_solver module
-----------------------------------------
-
-.. automodule:: src.numeric.solver.linear_solver
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-src.numeric.solver.nonlinear\_solver module
--------------------------------------------
-
-.. automodule:: src.numeric.solver.nonlinear_solver
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-src.numeric.solver.solver module
---------------------------------
-
-.. automodule:: src.numeric.solver.solver
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-src.numeric.solver.solvers module
----------------------------------
-
-.. automodule:: src.numeric.solver.solvers
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-src.numeric.solver.tunes module
--------------------------------
-
-.. automodule:: src.numeric.solver.tunes
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-src.numeric.solver.util module
-------------------------------
-
-.. automodule:: src.numeric.solver.util
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: src.numeric.solver
-   :members:
-   :undoc-members:
-   :show-inheritance:
