+++ supplements/api_docs/source/snowdrop.src.epidemic.rst
@@ -1,69 +1,0 @@
-snowdrop.src.epidemic package
-=============================
-
-Submodules
-----------
-
-snowdrop.src.epidemic.covid19 module
-------------------------------------
-
-.. automodule:: snowdrop.src.epidemic.covid19
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-snowdrop.src.epidemic.data module
----------------------------------
-
-.. automodule:: snowdrop.src.epidemic.data
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-snowdrop.src.epidemic.ert module
---------------------------------
-
-.. automodule:: snowdrop.src.epidemic.ert
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-snowdrop.src.epidemic.gsw module
---------------------------------
-
-.. automodule:: snowdrop.src.epidemic.gsw
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-snowdrop.src.epidemic.gsw\_tunes module
----------------------------------------
-
-.. automodule:: snowdrop.src.epidemic.gsw_tunes
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-snowdrop.src.epidemic.gsw\_v2 module
-------------------------------------
-
-.. automodule:: snowdrop.src.epidemic.gsw_v2
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-snowdrop.src.epidemic.sir\_params module
-----------------------------------------
-
-.. automodule:: snowdrop.src.epidemic.sir_params
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: snowdrop.src.epidemic
-   :members:
-   :undoc-members:
-   :show-inheritance:
