+++ supplements/api_docs/source/snowdrop.src.model.rst
@@ -1,53 +1,0 @@
-snowdrop.src.model package
-==========================
-
-Submodules
-----------
-
-snowdrop.src.model.factory module
----------------------------------
-
-.. automodule:: snowdrop.src.model.factory
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-snowdrop.src.model.interface module
------------------------------------
-
-.. automodule:: snowdrop.src.model.interface
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-snowdrop.src.model.model module
--------------------------------
-
-.. automodule:: snowdrop.src.model.model
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-snowdrop.src.model.settings module
-----------------------------------
-
-.. automodule:: snowdrop.src.model.settings
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-snowdrop.src.model.util module
-------------------------------
-
-.. automodule:: snowdrop.src.model.util
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: snowdrop.src.model
-   :members:
-   :undoc-members:
-   :show-inheritance:
