+++ supplements/api_docs/source/snowdrop.src.numeric.solver.rst
@@ -1,117 +1,0 @@
-snowdrop.src.numeric.solver package
-===================================
-
-Subpackages
------------
-
-.. toctree::
-   :maxdepth: 4
-
-   snowdrop.src.numeric.solver.AIM
-
-Submodules
-----------
-
-snowdrop.src.numeric.solver.ABLR module
----------------------------------------
-
-.. automodule:: snowdrop.src.numeric.solver.ABLR
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-snowdrop.src.numeric.solver.AndersonMoore module
-------------------------------------------------
-
-.. automodule:: snowdrop.src.numeric.solver.AndersonMoore
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-snowdrop.src.numeric.solver.Benes module
-----------------------------------------
-
-.. automodule:: snowdrop.src.numeric.solver.Benes
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-snowdrop.src.numeric.solver.BinderPesaran module
-------------------------------------------------
-
-.. automodule:: snowdrop.src.numeric.solver.BinderPesaran
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-snowdrop.src.numeric.solver.LBJ module
---------------------------------------
-
-.. automodule:: snowdrop.src.numeric.solver.LBJ
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-snowdrop.src.numeric.solver.Villemot module
--------------------------------------------
-
-.. automodule:: snowdrop.src.numeric.solver.Villemot
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-snowdrop.src.numeric.solver.linear\_solver module
--------------------------------------------------
-
-.. automodule:: snowdrop.src.numeric.solver.linear_solver
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-snowdrop.src.numeric.solver.nonlinear\_solver module
-----------------------------------------------------
-
-.. automodule:: snowdrop.src.numeric.solver.nonlinear_solver
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-snowdrop.src.numeric.solver.solver module
------------------------------------------
-
-.. automodule:: snowdrop.src.numeric.solver.solver
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-snowdrop.src.numeric.solver.solvers module
-------------------------------------------
-
-.. automodule:: snowdrop.src.numeric.solver.solvers
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-snowdrop.src.numeric.solver.tunes module
-----------------------------------------
-
-.. automodule:: snowdrop.src.numeric.solver.tunes
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-snowdrop.src.numeric.solver.util module
----------------------------------------
-
-.. automodule:: snowdrop.src.numeric.solver.util
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: snowdrop.src.numeric.solver
-   :members:
-   :undoc-members:
-   :show-inheritance:

[…]
