### 2025-08-02 d368398
Modified bibliography.
