+++ joss_submission/paper.bib
@@ -1,5 +1,5 @@
 @article{Eichenbaum, 
-  author = ({Eichenbaum, Martin and Rebelo, Sergio and Trabandt, Mathias }),
+  author = {Eichenbaum, Martin and Rebelo, Sergio and Trabandt, Mathias},
   journal = {NBER Working Paper, No. 26882},
   title = {The Macroeconomics of Epidemics},
   year = 2020,
@@ -9,7 +9,7 @@
 
 @article{fernandez,
   title = {Estimating DSGE models: Recent advances and future challenges},
-  author = ({Fern{\'a}ndez-Villaverde, Jes{\'u}s and Guerr{\'o}n-Quintana, Pablo A}),
+  author = {Fern{\'a}ndez-Villaverde, Jes{\'u}s and Guerr{\'o}n-Quintana, Pablo A},
   journal = {Annual Review of Economics},
   volume = {13},
   number = {1},
@@ -21,8 +21,8 @@
 }
 
 @article{botman,
-  title = ({DSGE modeling at the fund: Applications and further developments}),
-  author = ({Botman, Dennis PJ and Rose, David and Laxton, Douglas  and Karam, Philippe D}),
+  title = {DSGE modeling at the fund: Applications and further developments},
+  author = {Botman, Dennis PJ and Rose, David and Laxton, Douglas  and Karam, Philippe D},
   year = {2007},
   publisher = {IMF Working Paper},
   doi = {10.5089/9781451867640.001},
@@ -31,7 +31,7 @@
 
 @article{smets,
   title = {DSGE models and their use at the ECB},
-  author = ({Smets, Frank and Christoffel, Kai and Coenen, G{\"u}nter and Motto, Roberto and Rostagno, Massimo}),
+  author = {Smets, Frank and Christoffel, Kai and Coenen, G{\"u}nter and Motto, Roberto and Rostagno, Massimo},
   journal = {SERIEs},
   volume = {1},
   pages = {51--65},
@@ -43,7 +43,7 @@
 
 @article{del,
   title = {The FRBNY DSGE model},
-  author = ({Del Negro, Marco and Eusepi, Stefano and Giannoni, Marc P and Sbordone, Argia M and Tambalotti, Andrea and Cocci, Matthew and Hasegawa, Raiden and Linder, M}),
+  author = {Del Negro, Marco and Eusepi, Stefano and Giannoni, Marc P and Sbordone, Argia M and Tambalotti, Andrea and Cocci, Matthew and Hasegawa, Raiden and Linder, M},
   journal = {FRB of New York Staff Report},
   number = {647},
   year = {2013},
@@ -53,7 +53,7 @@
 
 @article{yagihashi,
   title = {DSGE models used by policymakers: A survey},
-  author = ({Yagihashi, Takeshi}),
+  author = {Yagihashi, Takeshi},
   journal = {Policy Research Institute, Ministry of Finance, Japan, PRI Discussion Paper Series},
   year = {2020},
   doi = {10.3390/econometrics4010001},
@@ -61,7 +61,7 @@
 }
 
 @article{Juillard, 
-  author = ({Juillard, Michel and Laxton, Douglas and McAdam, Peter and Pioro, Hope}),
+  author = {Juillard, Michel and Laxton, Douglas and McAdam, Peter and Pioro, Hope},
   journal = {Journal of Economic Dynamics and Control, No. 22, pp.1291—1318},
   title = {An algorithm competition: First-order iterations versus Newton-based technique},
   year = 1998,
@@ -70,7 +70,7 @@
 }
 
 @article{Goumilevski, 
-  author = ({Goumilevski Alexei and Farias Aquiles and Tanyeri Kadir}),
[…]
