+++ joss_submission/paper.bib
@@ -70,7 +70,7 @@
 }
 
 @article{Goumilevski, 
-  author = {Goumilevski Alexei and Farias Aquiles and Tanyeri Kadir},
+  author = {Goumilevski, Alexei and Farias, Aquiles and Tanyeri, Kadir},
   journal = {Proceedings of the 27th International Conference on Computing in Economics and Finance},
   title = {A New Macroeconomic Modeling Framework Applied to Assess Effects of COVID-19},
   year = 2021,
