### 2025-08-02 3bd50ea
Updated README file with the new release information.
