### 2025-08-03 #8197 @gumilevskij
Dear @Nikoleta-v3. Thanks a lot for your feedback.

I've made many changes to this GitHub repository.
In particular, I created a GitHub wiki page and uploaded the presentation slides.
I've published this software to the PYPI package index.

The Framework folder already has a snowdrop sub-folder.  I cannot upload the Framework
package to PYPI because a package with this name already exists there.

I also created a new release. The link to the Zenodo site is at the bottom of the README file.
Anyway, here's the info:

DOI: 10.5281/zenodo.16730040
URL: https://zenodo.org/records/16730040

Thanks, Alexei
