### 2025-08-03 a141dca
Minor fix to README
