### 2025-08-06 41e4d3a
Corrected bibliography.
