+++ joss_submission/paper (copy).bib
@@ -56,6 +56,7 @@
   author = {Yagihashi, Takeshi},
   journal = {Policy Research Institute, Ministry of Finance, Japan, PRI Discussion Paper Series},
   year = {2020},
+  doi = {10.3390/econometrics4010001},
   url = {https://www.mof.go.jp/pri/research/discussion_paper/ron333.pdf}
 }
 
