### 2025-08-06 4ba5642
Corrected bibliography.
