+++ joss_submission/paper (copy).bib
@@ -1,87 +1,0 @@
-@article{Eichenbaum, 
-  author = {Eichenbaum, Martin and Rebelo, Sergio and Trabandt, Mathias},
-  journal = {NBER Working Paper, No. 26882},
-  title = {The Macroeconomics of Epidemics},
-  year = 2020,
-  doi = {10.3386/w26882},
-  url = {https://www.nber.org/papers/w26882}
-}
-
-@article{fernandez,
-  title = {Estimating DSGE models: Recent advances and future challenges},
-  author = {Fern{\'a}ndez-Villaverde, Jes{\'u}s and Guerr{\'o}n-Quintana, Pablo A},
-  journal = {Annual Review of Economics},
-  volume = {13},
-  number = {1},
-  pages = {229--252},
-  year = {2021},
-  publisher = {Annual Reviews},
-  doi = {10.3386/w27715},
-  url = {https://www.nber.org/papers/w27715}
-}
-
-@article{botman,
-  title = {DSGE modeling at the fund: Applications and further developments},
-  author = {Botman, Dennis PJ and Rose, David and Laxton, Douglas  and Karam, Philippe D},
-  year = {2007},
-  publisher = {IMF Working Paper},
-  doi = {10.5089/9781451867640.001},
-  url = {https://www.imf.org/en/Publications/WP/Issues/2016/12/31/DSGE-Modeling-at-the-Fund-Applications-and-Further-Developments-21265}
-}
-
-@article{smets,
-  title = {DSGE models and their use at the ECB},
-  author = {Smets, Frank and Christoffel, Kai and Coenen, G{\"u}nter and Motto, Roberto and Rostagno, Massimo},
-  journal = {SERIEs},
-  volume = {1},
-  pages = {51--65},
-  year = {2010},
-  publisher = {Springer},
-  doi = {10.1007/s13209-010-0020-9},
-  url = {https://d-nb.info/1100877037/34}
-}
-
-@article{del,
-  title = {The FRBNY DSGE model},
-  author = {Del Negro, Marco and Eusepi, Stefano and Giannoni, Marc P and Sbordone, Argia M and Tambalotti, Andrea and Cocci, Matthew and Hasegawa, Raiden and Linder, M},
-  journal = {FRB of New York Staff Report},
-  number = {647},
-  year = {2013},
-  doi = {10.2139/ssrn.2344312},
-  url = {https://www.newyorkfed.org/medialibrary/media/research/staff_reports/sr647.pdf}
-}
-
-@article{yagihashi,
-  title = {DSGE models used by policymakers: A survey},
-  author = {Yagihashi, Takeshi},
-  journal = {Policy Research Institute, Ministry of Finance, Japan, PRI Discussion Paper Series},
-  year = {2020},
-  doi = {10.3390/econometrics4010001},
-  url = {https://www.mof.go.jp/pri/research/discussion_paper/ron333.pdf}
-}
-
-@article{Juillard, 
-  author = {Juillard, Michel and Laxton, Douglas and McAdam, Peter and Pioro, Hope},
-  journal = {Journal of Economic Dynamics and Control, No. 22, pp.1291—1318},
-  title = {An algorithm competition: First-order iterations versus Newton-based technique},
-  year = 1998,
-  doi = {10.1016/S0165-1889(98)00013-X},
-  url = {https://www.sciencedirect.com/science/article/abs/pii/S016518899800013X}
-}
-
-@article{Goumilevski, 
-  author = {Goumilevski, Alexei and Farias, Aquiles and Tanyeri, Kadir},
-  journal = {Proceedings of the 27th International Conference on Computing in Economics and Finance},
[…]
