### 2025-08-08 #8197 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
